+++ pkg/plugins/trivy/config_test.go
@@ -1,0 +1,818 @@
+package trivy
+
+import (
+	"context"
+	"time"
+
+	"testing"
+
+	"github.com/aquasecurity/trivy-operator/pkg/ext"
+	"github.com/aquasecurity/trivy-operator/pkg/kube"
+
+	"github.com/aquasecurity/trivy-operator/pkg/trivyoperator"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+
+	"sigs.k8s.io/controller-runtime/pkg/client/fake"
+)
+
+var (
+	fixedTime  = time.Now()
+	fixedClock = ext.NewFixedClock(fixedTime)
+)
+
+func TestConfig_GetImageRef(t *testing.T) {
+	testCases := []struct {
+		name             string
+		configData       Config
+		expectedError    string
+		expectedImageRef string
+	}{
+		{
+			name:          "Should return error",
+			configData:    Config{PluginConfig: trivyoperator.PluginConfig{}},
+			expectedError: "property repository not set",
+		},
+		{
+			name: "Should return error",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"tag": "0.8.0",
+				},
+			}},
+			expectedError: "property repository not set",
+		},
+		{
+			name: "Should return error",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"repository": "gcr.io/aquasecurity/trivy",
+				},
+			}},
+			expectedError: "property tag not set",
+		},
+		{
+			name: "Should return image reference from config data",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"repository": "gcr.io/aquasecurity/trivy",
+					"tag":        "0.8.0",
+				},
+			}},
+			expectedImageRef: "gcr.io/aquasecurity/trivy:0.8.0",
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			imageRef, err := tc.configData.GetImageRef()
+			if tc.expectedError != "" {
+				require.EqualError(t, err, tc.expectedError)
+			} else {
+				require.NoError(t, err)
+				assert.Equal(t, tc.expectedImageRef, imageRef)
+			}
+		})
+	}
+}
+
+func TestConfig_GetAdditionalVulnerabilityReportFields(t *testing.T) {
+	testCases := []struct {
+		name             string
+		configData       Config
+		additionalFields AdditionalFields
+	}{
+		{
+			name:             "no additional fields are set",
+			configData:       Config{PluginConfig: trivyoperator.PluginConfig{}},
+			additionalFields: AdditionalFields{},
+		},
+		{
+			name: "all additional fields are set",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"additionalVulnerabilityReportFields": "PackageType,PkgPath,Class,Target,Links,Description,CVSS",
+				},
+			}},
+			additionalFields: AdditionalFields{Description: true, Links: true, CVSS: true, Class: true, PackageType: true, PkgPath: true, Target: true},
+		},
+		{
+			name: "some additional fields are set",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"additionalVulnerabilityReportFields": "PackageType,Target,Links,CVSS",
+				},
+			}},
+			additionalFields: AdditionalFields{Links: true, CVSS: true, PackageType: true, Target: true},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			addFields := tc.configData.GetAdditionalVulnerabilityReportFields()
+			assert.True(t, addFields.Description == tc.additionalFields.Description)
+			assert.True(t, addFields.CVSS == tc.additionalFields.CVSS)
+			assert.True(t, addFields.Target == tc.additionalFields.Target)
+			assert.True(t, addFields.PackageType == tc.additionalFields.PackageType)
+			assert.True(t, addFields.Class == tc.additionalFields.Class)
+			assert.True(t, addFields.Links == tc.additionalFields.Links)
+		})
+	}
+}
+
+func TestConfig_GetMode(t *testing.T) {
+	testCases := []struct {
+		name          string
+		configData    Config
+		expectedError string
+		expectedMode  Mode
+	}{
+		{
+			name: "Should return Standalone",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"mode": string(Standalone),
+				},
+			}},
+			expectedMode: Standalone,
+		},
+		{
+			name: "Should return ClientServer",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"mode": string(ClientServer),
+				},
+			}},
+			expectedMode: ClientServer,
+		},
+		{
+			name:          "Should return error when value is not set",
+			configData:    Config{PluginConfig: trivyoperator.PluginConfig{}},
+			expectedError: "property mode not set",
+		},
+		{
+			name: "Should return error when value is not allowed",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"mode": "P2P",
+				},
+			}},
+			expectedError: "invalid value (P2P) of mode; allowed values (Standalone, ClientServer)",
+		},
+	}
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			mode, err := tc.configData.GetMode()
+			if tc.expectedError != "" {
+				require.EqualError(t, err, tc.expectedError)
+			} else {
+				require.NoError(t, err)
+				assert.Equal(t, tc.expectedMode, mode)
+			}
+		})
+	}
+}
+
+func TestGetSlow(t *testing.T) {
+	testCases := []struct {
+		name       string
+		configData Config
+		want       bool
+	}{
+		{
+			name: "slow param set to true",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"slow": "true",
+				},
+			}},
+			want: true,
+		},
+		{
+			name: "slow param set to false",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"slow": "false",
+				},
+			}},
+			want: false,
+		},
+		{
+			name: "slow param set to no valid value",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"slow": "false2",
+				},
+			}},
+			want: true,
+		},
+		{
+			name: "slow param set to no  value",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{},
+			}},
+			want: true,
+		},
+	}
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			got := tc.configData.GetSlow()
+			assert.Equal(t, got, tc.want)
+
+		})
+	}
+}
+func TestConfig_GetCommand(t *testing.T) {
+	testCases := []struct {
+		name            string
+		configData      Config
+		expectedError   string
+		expectedCommand Command
+	}{
+		{
+			name: "Should return image",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"command": "image",
+				},
+			}},
+			expectedCommand: Image,
+		},
+		{
+			name: "Should return image when value is not set",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{},
+			}},
+			expectedCommand: Image,
+		},
+		{
+			name: "Should return filesystem",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"command": "filesystem",
+				},
+			}},
+			expectedCommand: Filesystem,
+		},
+		{
+			name: "Should return rootfs",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"command": "rootfs",
+				},
+			}},
+			expectedCommand: Rootfs,
+		},
+		{
+			name: "Should return error when value is not allowed",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"command": "ls",
+				},
+			}},
+			expectedError: "invalid value (ls) of command; allowed values (image, filesystem, rootfs)",
+		},
+	}
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			command, err := tc.configData.GetCommand()
+			if tc.expectedError != "" {
+				require.EqualError(t, err, tc.expectedError)
+			} else {
+				require.NoError(t, err)
+				assert.Equal(t, tc.expectedCommand, command)
+			}
+		})
+	}
+}
+
+func TestVulnType(t *testing.T) {
+	testCases := []struct {
+		name       string
+		configData Config
+		want       string
+	}{
+		{
+			name: "valid vuln type os",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"vulnType": "os",
+				},
+			}},
+			want: "os",
+		},
+		{
+			name: "valid vuln type library",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"vulnType": "library",
+				},
+			}},
+			want: "library",
+		},
+		{
+			name: "empty vuln type",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"vulnType": "",
+				},
+			}},
+			want: "",
+		},
+		{
+			name: "non valid vuln type",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"vulnType": "aaa",
+				},
+			}},
+			want: "",
+		},
+	}
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			got := tc.configData.GetVulnType()
+			assert.Equal(t, got, tc.want)
+
+		})
+	}
+}
+
+func TestConfig_GetResourceRequirements(t *testing.T) {
+	testCases := []struct {
+		name                 string
+		config               Config
+		expectedError        string
+		expectedRequirements corev1.ResourceRequirements
+	}{
+		{
+			name: "Should return empty requirements by default",
+			config: Config{
+				PluginConfig: trivyoperator.PluginConfig{},
+			},
+			expectedError: "",
+			expectedRequirements: corev1.ResourceRequirements{
+				Requests: corev1.ResourceList{},
+				Limits:   corev1.ResourceList{},
+			},
+		},
+		{
+			name: "Should return configured resource requirement",
+			config: Config{
+				PluginConfig: trivyoperator.PluginConfig{
+					Data: map[string]string{
+						"dbRepository":              DefaultDBRepository,
+						"javaDbRepository":          DefaultJavaDBRepository,
+						"resources.requests.cpu":    "800m",
+						"resources.requests.memory": "200M",
+						"resources.limits.cpu":      "600m",
+						"resources.limits.memory":   "700M",
+					},
+				},
+			},
+			expectedError: "",
+			expectedRequirements: corev1.ResourceRequirements{
+				Requests: corev1.ResourceList{
+					corev1.ResourceCPU:    resource.MustParse("800m"),
+					corev1.ResourceMemory: resource.MustParse("200M"),
+				},
+				Limits: corev1.ResourceList{
+					corev1.ResourceCPU:    resource.MustParse("600m"),
+					corev1.ResourceMemory: resource.MustParse("700M"),
+				},
+			},
+		},
+		{
+			name: "Should return error if resource is not parseable",
+			config: Config{
+				PluginConfig: trivyoperator.PluginConfig{
+					Data: map[string]string{
+						"resources.requests.cpu": "roughly 100",
+					},
+				},
+			},
+			expectedError: "parsing resource definition resources.requests.cpu: roughly 100 quantities must match the regular expression '^([+-]?[0-9.]+)([eEinumkKMGTP]*[-+]?[0-9]*)$'",
+		},
+	}
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			resourceRequirement, err := tc.config.GetResourceRequirements()
+			if tc.expectedError != "" {
+				require.EqualError(t, err, tc.expectedError)
+			} else {
+				require.NoError(t, err)
+				assert.Equal(t, tc.expectedRequirements, resourceRequirement, tc.name)
+			}
+		})
+	}
+}
+
+func TestConfig_IgnoreFileExists(t *testing.T) {
+	testCases := []struct {
+		name           string
+		configData     Config
+		expectedOutput bool
+	}{
+		{
+			name: "Should return false",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"foo": "bar",
+				},
+			}},
+			expectedOutput: false,
+		},
+		{
+			name: "Should return true",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"foo": "bar",
+					"ignoreFile": `# Accept the risk
+CVE-2018-14618
+
+# No impact in our settings
+CVE-2019-1543`,
+				},
+			}},
+			expectedOutput: true,
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			exists := tc.configData.IgnoreFileExists()
+			assert.Equal(t, tc.expectedOutput, exists)
+		})
+	}
+}
+
+func TestConfig_IgnoreUnfixed(t *testing.T) {
+	testCases := []struct {
+		name           string
+		configData     Config
+		expectedOutput bool
+	}{
+		{
+			name: "Should return false",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"foo": "bar",
+				},
+			}},
+			expectedOutput: false,
+		},
+		{
+			name: "Should return true",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"foo":           "bar",
+					"ignoreUnfixed": "true",
+				},
+			}},
+			expectedOutput: true,
+		},
+		{
+			name: "Should return false when set it as false",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"foo":           "bar",
+					"ignoreUnfixed": "false",
+				},
+			}},
+			expectedOutput: true,
+		},
+	}
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			exists := tc.configData.IgnoreUnfixed()
+			assert.Equal(t, tc.expectedOutput, exists)
+		})
+	}
+}
+
+func TestConfig_OfflineScan(t *testing.T) {
+	testCases := []struct {
+		name           string
+		configData     Config
+		expectedOutput bool
+	}{
+		{
+			name: "Should return false",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"foo": "bar",
+				},
+			}},
+			expectedOutput: false,
+		},
+		{
+			name: "Should return true",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"foo":         "bar",
+					"offlineScan": "true",
+				},
+			}},
+			expectedOutput: true,
+		},
+		{
+			name: "Should return false when set it as false",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"foo":         "bar",
+					"offlineScan": "false",
+				},
+			}},
+			expectedOutput: true,
+		},
+	}
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			exists := tc.configData.OfflineScan()
+			assert.Equal(t, tc.expectedOutput, exists)
+		})
+	}
+}
+
+func TestConfig_dbRepositoryInsecure(t *testing.T) {
+	testCases := []struct {
+		name           string
+		configData     Config
+		expectedOutput bool
+	}{
+		{
+			name: "good value Should return false",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"dbRepositoryInsecure": "false",
+				},
+			}},
+			expectedOutput: false,
+		},
+		{
+			name: "good value Should return true",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"dbRepositoryInsecure": "true",
+				},
+			}},
+			expectedOutput: true,
+		},
+		{
+			name: "bad value Should return false",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"dbRepositoryInsecure": "true1",
+				},
+			}},
+			expectedOutput: false,
+		},
+		{
+			name: "no value Should return false",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{},
+			}},
+			expectedOutput: false,
+		},
+	}
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			exists := tc.configData.GetDBRepositoryInsecure()
+			assert.Equal(t, tc.expectedOutput, exists)
+		})
+	}
+}
+
+func TestConfig_GetInsecureRegistries(t *testing.T) {
+	testCases := []struct {
+		name           string
+		configData     Config
+		expectedOutput map[string]bool
+	}{
+		{
+			name: "Should return nil map when there is no key with insecureRegistry. prefix",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"foo": "bar",
+				},
+			}},
+			expectedOutput: make(map[string]bool),
+		},
+		{
+			name: "Should return insecure registries in map",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"foo":                          "bar",
+					"insecureRegistry.pocRegistry": "poc.myregistry.harbor.com.pl",
+					"insecureRegistry.qaRegistry":  "qa.registry.aquasec.com",
+				},
+			}},
+			expectedOutput: map[string]bool{
+				"poc.myregistry.harbor.com.pl": true,
+				"qa.registry.aquasec.com":      true,
+			},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			insecureRegistries := tc.configData.GetInsecureRegistries()
+			assert.Equal(t, tc.expectedOutput, insecureRegistries)
+		})
+	}
+}
+
+func TestConfig_GetNonSSLRegistries(t *testing.T) {
+	testCases := []struct {
+		name           string
+		configData     Config
+		expectedOutput map[string]bool
+	}{
+		{
+			name: "Should return nil map when there is no key with nonSslRegistry. prefix",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"foo": "bar",
+				},
+			}},
+			expectedOutput: make(map[string]bool),
+		},
+		{
+			name: "Should return insecure registries in map",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"foo":                        "bar",
+					"nonSslRegistry.pocRegistry": "poc.myregistry.harbor.com.pl",
+					"nonSslRegistry.qaRegistry":  "qa.registry.aquasec.com",
+				},
+			}},
+			expectedOutput: map[string]bool{
+				"poc.myregistry.harbor.com.pl": true,
+				"qa.registry.aquasec.com":      true,
+			},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			nonSslRegistries := tc.configData.GetNonSSLRegistries()
+			assert.Equal(t, tc.expectedOutput, nonSslRegistries)
+		})
+	}
+}
+
+func TestConfig_GetMirrors(t *testing.T) {
+	testCases := []struct {
+		name           string
+		configData     Config
+		expectedOutput map[string]string
+	}{
+		{
+			name: "Should return empty map when there is no key with mirrors.registry. prefix",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"foo": "bar",
+				},
+			}},
+			expectedOutput: make(map[string]string),
+		},
+		{
+			name: "Should return mirrors in a map",
+			configData: Config{PluginConfig: trivyoperator.PluginConfig{
+				Data: map[string]string{
+					"registry.mirror.docker.io": "mirror.io",
+				},
+			}},
+			expectedOutput: map[string]string{"docker.io": "mirror.io"},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			assert.Equal(t, tc.expectedOutput, tc.configData.GetMirrors())
+		})
+	}
+}
+
+func TestPlugin_Init(t *testing.T) {
+
+	t.Run("Should create the default config", func(t *testing.T) {
+		testClient := fake.NewClientBuilder().WithObjects().Build()
+		or := kube.NewObjectResolver(testClient, &kube.CompatibleObjectMapper{})
+
+		pluginContext := trivyoperator.NewPluginContext().
+			WithName(Plugin).
+			WithNamespace("trivyoperator-ns").
+			WithServiceAccountName("trivyoperator-sa").
+			WithClient(testClient).
+			Get()
+		psm, err := NewPodSpecMgr(pluginContext)
+		assert.NoError(t, err)
+		instance := NewPlugin(fixedClock, ext.NewSimpleIDGenerator(), &or, psm)
+
+		err = instance.Init(pluginContext)
+		require.NoError(t, err)
+
+		var cm corev1.ConfigMap
+		err = testClient.Get(context.Background(), types.NamespacedName{
+			Namespace: "trivyoperator-ns",
+			Name:      "trivy-operator-trivy-config",
+		}, &cm)
+		require.NoError(t, err)
+		assert.Equal(t, corev1.ConfigMap{
+			TypeMeta: metav1.TypeMeta{
+				APIVersion: "v1",
+				Kind:       "ConfigMap",
+			},
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "trivy-operator-trivy-config",
+				Namespace: "trivyoperator-ns",
+				Labels: map[string]string{
+					"app.kubernetes.io/managed-by": "trivyoperator",
+				},
+				ResourceVersion: "1",
+			},
+			Data: map[string]string{
+				"repository":                DefaultImageRepository,
+				"tag":                       "0.45.1",
+				"severity":                  DefaultSeverity,
+				"slow":                      "true",
+				"mode":                      string(Standalone),
+				"timeout":                   "5m0s",
+				"dbRepository":              DefaultDBRepository,
+				"javaDbRepository":          DefaultJavaDBRepository,
+				"useBuiltinRegoPolicies":    "true",
+				"supportedConfigAuditKinds": SupportedConfigAuditKinds,
+				"resources.requests.cpu":    "100m",
+				"resources.requests.memory": "100M",
+				"resources.limits.cpu":      "500m",
+				"resources.limits.memory":   "500M",
+			},
+		}, cm)
+	})
+
+	t.Run("Should not overwrite existing config", func(t *testing.T) {
+		testClient := fake.NewClientBuilder().WithObjects(
+			&corev1.ConfigMap{
+				TypeMeta: metav1.TypeMeta{
+					APIVersion: "v1",
+					Kind:       "ConfigMap",
+				},
+				ObjectMeta: metav1.ObjectMeta{
+					Name:            "trivy-operator-trivy-config",
+					Namespace:       "trivyoperator-ns",
+					ResourceVersion: "1",
+				},
+				Data: map[string]string{
+					"repository": "gcr.io/aquasecurity/trivy",
+					"tag":        "0.35.0",
+					"severity":   DefaultSeverity,
+					"mode":       string(Standalone),
+				},
+			}).Build()
+		resolver := kube.NewObjectResolver(testClient, &kube.CompatibleObjectMapper{})
+
+		pluginContext := trivyoperator.NewPluginContext().
+			WithName(Plugin).
+			WithNamespace("trivyoperator-ns").
+			WithServiceAccountName("trivyoperator-sa").
+			WithClient(testClient).
+			Get()
+
+		psm, err := NewPodSpecMgr(pluginContext)
+		assert.NoError(t, err)
+		instance := NewPlugin(fixedClock, ext.NewSimpleIDGenerator(), &resolver, psm)
+		err = instance.Init(pluginContext)
+		require.NoError(t, err)
+
+		var cm corev1.ConfigMap
+		err = testClient.Get(context.Background(), types.NamespacedName{
+			Namespace: "trivyoperator-ns",
+			Name:      "trivy-operator-trivy-config",
+		}, &cm)
+		require.NoError(t, err)
+		assert.Equal(t, corev1.ConfigMap{
+			TypeMeta: metav1.TypeMeta{
+				APIVersion: "v1",
+				Kind:       "ConfigMap",
+			},
+			ObjectMeta: metav1.ObjectMeta{
+				Name:            "trivy-operator-trivy-config",
+				Namespace:       "trivyoperator-ns",
+				ResourceVersion: "1",
+			},
+			Data: map[string]string{
+				"repository": "gcr.io/aquasecurity/trivy",
+				"tag":        "0.35.0",
+				"severity":   DefaultSeverity,
+				"mode":       string(Standalone),
+			},
+		}, cm)
+	})
+}
