+++ pkg/plugins/trivy/plugin.go
@@ -1,14 +1,12 @@
 package trivy
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"io"
-	"net/url"
-	"path/filepath"
+
 	"regexp"
-	"strconv"
+
 	"strings"
 	"time"
 
@@ -29,7 +27,7 @@ import (
 	"github.com/aquasecurity/trivy-operator/pkg/trivyoperator"
 	"github.com/aquasecurity/trivy-operator/pkg/vulnerabilityreport"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
+
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/utils/pointer"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -48,59 +46,6 @@ const (
 	SkipDirsAnnotation = "trivy-operator.aquasecurity.github.io/skip-dirs"
 	// SkipFilesAnnotation example: trivy-operator.aquasecurity.github.io/skip-files: "/src/Gemfile.lock,/examplebinary"
 	SkipFilesAnnotation = "trivy-operator.aquasecurity.github.io/skip-files"
-)
-
-const (
-	keyTrivyImageRepository = "trivy.repository"
-	keyTrivyImageTag        = "trivy.tag"
-	//nolint:gosec
-	keyTrivyImagePullSecret                     = "trivy.imagePullSecret"
-	keyTrivyImagePullPolicy                     = "trivy.imagePullPolicy"
-	keyTrivyMode                                = "trivy.mode"
-	keyTrivyAdditionalVulnerabilityReportFields = "trivy.additionalVulnerabilityReportFields"
-	keyTrivyCommand                             = "trivy.command"
-	KeyTrivySeverity                            = "trivy.severity"
-	keyTrivySlow                                = "trivy.slow"
-	keyTrivyVulnType                            = "trivy.vulnType"
-	keyTrivyIgnoreUnfixed                       = "trivy.ignoreUnfixed"
-	keyTrivyOfflineScan                         = "trivy.offlineScan"
-	keyTrivyTimeout                             = "trivy.timeout"
-	keyTrivyIgnoreFile                          = "trivy.ignoreFile"
-	keyTrivyIgnorePolicy                        = "trivy.ignorePolicy"
-	keyTrivyInsecureRegistryPrefix              = "trivy.insecureRegistry."
-	keyTrivyNonSslRegistryPrefix                = "trivy.nonSslRegistry."
-	keyTrivyMirrorPrefix                        = "trivy.registry.mirror."
-	keyTrivyHTTPProxy                           = "trivy.httpProxy"
-	keyTrivyHTTPSProxy                          = "trivy.httpsProxy"
-	keyTrivyNoProxy                             = "trivy.noProxy"
-	keyTrivySslCertDir                          = "trivy.sslCertDir"
-	// nolint:gosec // This is not a secret, but a configuration value.
-	keyTrivyGitHubToken          = "trivy.githubToken"
-	keyTrivySkipFiles            = "trivy.skipFiles"
-	keyTrivySkipDirs             = "trivy.skipDirs"
-	keyTrivyDBRepository         = "trivy.dbRepository"
-	keyTrivyJavaDBRepository     = "trivy.javaDbRepository"
-	keyTrivyDBRepositoryInsecure = "trivy.dbRepositoryInsecure"
-
-	keyTrivyUseBuiltinRegoPolicies    = "trivy.useBuiltinRegoPolicies"
-	keyTrivySupportedConfigAuditKinds = "trivy.supportedConfigAuditKinds"
-
-	keyTrivyServerURL              = "trivy.serverURL"
-	keyTrivyClientServerSkipUpdate = "trivy.clientServerSkipUpdate"
-	keyTrivySkipJavaDBUpdate       = "trivy.skipJavaDBUpdate"
-	// nolint:gosec // This is not a secret, but a configuration value.
-	keyTrivyServerTokenHeader = "trivy.serverTokenHeader"
-	keyTrivyServerInsecure    = "trivy.serverInsecure"
-	// nolint:gosec // This is not a secret, but a configuration value.
-	keyTrivyServerToken         = "trivy.serverToken"
-	keyTrivyServerCustomHeaders = "trivy.serverCustomHeaders"
-
-	keyResourcesRequestsCPU             = "trivy.resources.requests.cpu"
-	keyResourcesRequestsMemory          = "trivy.resources.requests.memory"
-	keyResourcesLimitsCPU               = "trivy.resources.limits.cpu"
-	keyResourcesLimitsMemory            = "trivy.resources.limits.memory"
-	keyResourcesRequestEphemeralStorage = "trivy.resources.requests.ephemeral-storage"
-	keyResourcesLimitEphemeralStorage   = "trivy.resources.limits.ephemeral-storage"
 )
 
 const (
@@ -137,425 +82,11 @@ type AdditionalFields struct {
 	PkgPath     bool
 }
 
-// Config defines configuration params for this plugin.
-type Config struct {
-	trivyoperator.PluginConfig
-}
-
-func (c Config) GetAdditionalVulnerabilityReportFields() AdditionalFields {
-	addFields := AdditionalFields{}
-
-	fields, ok := c.Data[keyTrivyAdditionalVulnerabilityReportFields]
-	if !ok {
-		return addFields
-	}
-	for _, field := range strings.Split(fields, ",") {
-		switch strings.TrimSpace(field) {
-		case "Description":
-			addFields.Description = true
-		case "Links":
-			addFields.Links = true
-		case "CVSS":
-			addFields.CVSS = true
-		case "Target":
-			addFields.Target = true
-		case "Class":
-			addFields.Class = true
-		case "PackageType":
-			addFields.PackageType = true
-		case "PackagePath":
-			addFields.PkgPath = true
-		}
-	}
-	return addFields
-}
-
-// GetImageRef returns upstream Trivy container image reference.
-func (c Config) GetImageRef() (string, error) {
-	repository, err := c.GetRequiredData(keyTrivyImageRepository)
-	if err != nil {
-		return "", err
-	}
-	tag, err := c.GetImageTag()
-	if err != nil {
-		return "", err
-	}
-
-	return fmt.Sprintf("%s:%s", repository, tag), nil
-}
-
-// GetImageTag returns upstream Trivy container image tag.
-func (c Config) GetImageTag() (string, error) {
-	tag, err := c.GetRequiredData(keyTrivyImageTag)
-	if err != nil {
-		return "", err
-	}
-	return tag, nil
-}
-
-func (c Config) GetImagePullSecret() []corev1.LocalObjectReference {
-	ips, ok := c.Data[keyTrivyImagePullSecret]
-	if !ok {
-		return []corev1.LocalObjectReference{}
-	}
-	return []corev1.LocalObjectReference{{Name: ips}}
-}
-
-func (c Config) GetImagePullPolicy() string {
-	ipp, ok := c.Data[keyTrivyImagePullPolicy]
-	if !ok {
-		return "IfNotPresent"
-	}
-	return ipp
-}
-
-func (c Config) GetMode() (Mode, error) {
-	var ok bool
-	var value string
-	if value, ok = c.Data[keyTrivyMode]; !ok {
-		return "", fmt.Errorf("property %s not set", keyTrivyMode)
-	}
-
-	switch Mode(value) {
-	case Standalone:
-		return Standalone, nil
-	case ClientServer:
-		return ClientServer, nil
-	}
-
-	return "", fmt.Errorf("invalid value (%s) of %s; allowed values (%s, %s)",
-		value, keyTrivyMode, Standalone, ClientServer)
-}
-
-func (c Config) GetCommand() (Command, error) {
-	var ok bool
-	var value string
-	if value, ok = c.Data[keyTrivyCommand]; !ok {
-		// for backward compatibility, fallback to ImageScan
-		return Image, nil
-	}
-	switch Command(value) {
-	case Image:
-		return Image, nil
-	case Filesystem:
-		return Filesystem, nil
-	case Rootfs:
-		return Rootfs, nil
-	}
-	return "", fmt.Errorf("invalid value (%s) of %s; allowed values (%s, %s, %s)",
-		value, keyTrivyCommand, Image, Filesystem, Rootfs)
-}
-
-func (c Config) GetServerURL() (string, error) {
-	return c.GetRequiredData(keyTrivyServerURL)
-}
-
-func (c Config) GetClientServerSkipUpdate() bool {
-	val, ok := c.Data[keyTrivyClientServerSkipUpdate]
-	if !ok {
-		return false
-	}
-	boolVal, err := strconv.ParseBool(val)
-	if err != nil {
-		return false
-	}
-	return boolVal
-}
-
-func (c Config) GetSkipJavaDBUpdate() bool {
-	val, ok := c.Data[keyTrivySkipJavaDBUpdate]
-	if !ok {
-		return false
-	}
-	boolVal, err := strconv.ParseBool(val)
-	if err != nil {
-		return false
-	}
-	return boolVal
-}
-
-func (c Config) GetServerInsecure() bool {
-	_, ok := c.Data[keyTrivyServerInsecure]
-	return ok
-}
-
-func (c Config) GetDBRepositoryInsecure() bool {
-	val, ok := c.Data[keyTrivyDBRepositoryInsecure]
-	if !ok {
-		return false
-	}
-	boolVal, _ := strconv.ParseBool(val)
-	return boolVal
-}
-func (c Config) GetUseBuiltinRegoPolicies() bool {
-	val, ok := c.Data[keyTrivyUseBuiltinRegoPolicies]
-	if !ok {
-		return true
-	}
-	boolVal, err := strconv.ParseBool(val)
-	if err != nil {
-		return true
-	}
-	return boolVal
-}
-func (c Config) GetSslCertDir() string {
-	val, ok := c.Data[keyTrivySslCertDir]
-	if !ok {
-		return ""
-	}
-	return val
-}
-
-func (c Config) GetSeverity() string {
-	val, ok := c.Data[KeyTrivySeverity]
-	if !ok {
-		return ""
-	}
-	return val
-}
-
-func (c Config) GetSlow() bool {
-	val, ok := c.Data[keyTrivySlow]
-	if !ok {
-		return true
-	}
-	boolVal, err := strconv.ParseBool(val)
-	if err != nil {
-		return true
-	}
-	return boolVal
-}
-
-func (c Config) GetVulnType() string {
-	val, ok := c.Data[keyTrivyVulnType]
-	if !ok {
-		return ""
-	}
-	trimmedVulnType := strings.TrimSpace(val)
-	if !(trimmedVulnType == "os" || trimmedVulnType == "library") {
-		return ""
-	}
-	return trimmedVulnType
-}
-
-func (c Config) GetSupportedConfigAuditKinds() []string {
-	val, ok := c.Data[keyTrivySupportedConfigAuditKinds]
-	if !ok {
-		return utils.MapKinds(strings.Split(SupportedConfigAuditKinds, ","))
-	}
-	return utils.MapKinds(strings.Split(val, ","))
-}
-
-func (c Config) IgnoreFileExists() bool {
-	_, ok := c.Data[keyTrivyIgnoreFile]
-	return ok
-}
-
-func (c Config) FindIgnorePolicyKey(workload client.Object) string {
-	keysByPrecedence := []string{
-		keyTrivyIgnorePolicy + "." + workload.GetNamespace() + "." + workload.GetName(),
-		keyTrivyIgnorePolicy + "." + workload.GetNamespace(),
-		keyTrivyIgnorePolicy,
-	}
-	for _, key := range keysByPrecedence {
-		for key2 := range c.Data {
-			if key2 == keyTrivyIgnorePolicy || strings.HasPrefix(key2, keyTrivyIgnorePolicy) {
-				tempKey := key2
-				if key2 != keyTrivyIgnorePolicy {
-					// replace dot with astrix for regex matching
-					tempKey = fmt.Sprintf("%s%s", keyTrivyIgnorePolicy, strings.ReplaceAll(tempKey[len(keyTrivyIgnorePolicy):], ".", "*"))
-				}
-				matched, err := filepath.Match(tempKey, key)
-				if err == nil && matched {
-					return key2
-				}
-			}
-		}
-	}
-	return ""
-}
-
-func (c Config) GenerateIgnoreFileVolumeIfAvailable(trivyConfigName string) (*corev1.Volume, *corev1.VolumeMount) {
-	if !c.IgnoreFileExists() {
-		return nil, nil
-	}
-	volume := corev1.Volume{
-		Name: ignoreFileVolumeName,
-		VolumeSource: corev1.VolumeSource{
-			ConfigMap: &corev1.ConfigMapVolumeSource{
-				LocalObjectReference: corev1.LocalObjectReference{
-					Name: trivyConfigName,
-				},
-				Items: []corev1.KeyToPath{
-					{
-						Key:  keyTrivyIgnoreFile,
-						Path: ignoreFileName,
-					},
-				},
-			},
-		},
-	}
-	volumeMount := corev1.VolumeMount{
-		Name:      ignoreFileVolumeName,
-		MountPath: ignoreFileMountPath,
-		SubPath:   ignoreFileName,
-	}
-	return &volume, &volumeMount
-}
-
-func (c Config) GenerateSslCertDirVolumeIfAvailable(trivyConfigName string) (*corev1.Volume, *corev1.VolumeMount) {
-	var sslCertDirHost string
-	if sslCertDirHost = c.GetSslCertDir(); len(sslCertDirHost) == 0 {
-		return nil, nil
-	}
-	volume := corev1.Volume{
-		Name: sslCertDirVolumeName,
-		VolumeSource: corev1.VolumeSource{
-			HostPath: &corev1.HostPathVolumeSource{
-				Path: sslCertDirHost,
-			},
-		},
-	}
-	volumeMount := corev1.VolumeMount{
-		Name:      sslCertDirVolumeName,
-		MountPath: SslCertDir,
-		ReadOnly:  true,
-	}
-	return &volume, &volumeMount
-}
-
-func (c Config) GenerateIgnorePolicyVolumeIfAvailable(trivyConfigName string, workload client.Object) (*corev1.Volume, *corev1.VolumeMount) {
-	ignorePolicyKey := c.FindIgnorePolicyKey(workload)
-	if ignorePolicyKey == "" {
-		return nil, nil
-	}
-	volume := corev1.Volume{
-		Name: ignorePolicyVolumeName,
-		VolumeSource: corev1.VolumeSource{
-			ConfigMap: &corev1.ConfigMapVolumeSource{
-				LocalObjectReference: corev1.LocalObjectReference{
-					Name: trivyConfigName,
-				},
-				Items: []corev1.KeyToPath{
-					{
-						Key:  c.FindIgnorePolicyKey(workload),
-						Path: ignorePolicyName,
-					},
-				},
-			},
-		},
-	}
-	volumeMounts := corev1.VolumeMount{
-		Name:      ignorePolicyVolumeName,
-		MountPath: ignorePolicyMountPath,
-		SubPath:   ignorePolicyName,
-	}
-	return &volume, &volumeMounts
-}
-
-func (c Config) IgnoreUnfixed() bool {
-	_, ok := c.Data[keyTrivyIgnoreUnfixed]
-	return ok
-}
-
-func (c Config) OfflineScan() bool {
-	_, ok := c.Data[keyTrivyOfflineScan]
-	return ok
-}
-
-func (c Config) GetInsecureRegistries() map[string]bool {
-	insecureRegistries := make(map[string]bool)
-	for key, val := range c.Data {
-		if strings.HasPrefix(key, keyTrivyInsecureRegistryPrefix) {
-			insecureRegistries[val] = true
-		}
-	}
-
-	return insecureRegistries
-}
-
-func (c Config) GetNonSSLRegistries() map[string]bool {
-	nonSSLRegistries := make(map[string]bool)
-	for key, val := range c.Data {
-		if strings.HasPrefix(key, keyTrivyNonSslRegistryPrefix) {
-			nonSSLRegistries[val] = true
-		}
-	}
-
-	return nonSSLRegistries
-}
-
-func (c Config) GetMirrors() map[string]string {
-	res := make(map[string]string)
-	for registryKey, mirror := range c.Data {
-		if !strings.HasPrefix(registryKey, keyTrivyMirrorPrefix) {
-			continue
-		}
-		res[strings.TrimPrefix(registryKey, keyTrivyMirrorPrefix)] = mirror
-	}
-	return res
-}
-
-// GetResourceRequirements creates ResourceRequirements from the Config.
-func (c Config) GetResourceRequirements() (corev1.ResourceRequirements, error) {
-	requirements := corev1.ResourceRequirements{
-		Requests: corev1.ResourceList{},
-		Limits:   corev1.ResourceList{},
-	}
-
-	err := c.setResourceLimit(keyResourcesRequestsCPU, &requirements.Requests, corev1.ResourceCPU)
-	if err != nil {
-		return requirements, err
-	}
-
-	err = c.setResourceLimit(keyResourcesRequestsMemory, &requirements.Requests, corev1.ResourceMemory)
-	if err != nil {
-		return requirements, err
-	}
-
-	err = c.setResourceLimit(keyResourcesRequestEphemeralStorage, &requirements.Requests, corev1.ResourceEphemeralStorage)
-	if err != nil {
-		return requirements, err
-	}
-
-	err = c.setResourceLimit(keyResourcesLimitsCPU, &requirements.Limits, corev1.ResourceCPU)
-	if err != nil {
-		return requirements, err
-	}
-
-	err = c.setResourceLimit(keyResourcesLimitsMemory, &requirements.Limits, corev1.ResourceMemory)
-	if err != nil {
-		return requirements, err
-	}
-
-	err = c.setResourceLimit(keyResourcesLimitEphemeralStorage, &requirements.Limits, corev1.ResourceEphemeralStorage)
-	if err != nil {
-		return requirements, err
-	}
-
-	return requirements, nil
-}
-
-func (c Config) setResourceLimit(configKey string, k8sResourceList *corev1.ResourceList, k8sResourceName corev1.ResourceName) error {
-	if value, found := c.Data[configKey]; found {
-		quantity, err := resource.ParseQuantity(value)
-		if err != nil {
-			return fmt.Errorf("parsing resource definition %s: %s %w", configKey, value, err)
-		}
-
-		(*k8sResourceList)[k8sResourceName] = quantity
-	}
-	return nil
-}
-
-func (c Config) GetDBRepository() (string, error) {
-	return c.GetRequiredData(keyTrivyDBRepository)
-}
-
 type plugin struct {
 	clock          ext.Clock
 	idGenerator    ext.IDGenerator
 	objectResolver *kube.ObjectResolver
+	podSpecMgr     PodSpecMgr
 }
 
 // NewPlugin constructs a new vulnerabilityreport.Plugin, which is using an
@@ -569,11 +100,12 @@ type plugin struct {
 // on the settings returned by Config.GetMode. The ClientServer mode is usually
 // more performant, however it requires a Trivy server accessible at the
 // configurable Config.GetServerURL.
-func NewPlugin(clock ext.Clock, idGenerator ext.IDGenerator, objectResolver *kube.ObjectResolver) vulnerabilityreport.Plugin {
+func NewPlugin(clock ext.Clock, idGenerator ext.IDGenerator, objectResolver *kube.ObjectResolver, podSpecMgr PodSpecMgr) vulnerabilityreport.Plugin {
 	return &plugin{
 		clock:          clock,
 		idGenerator:    idGenerator,
 		objectResolver: objectResolver,
+		podSpecMgr:     podSpecMgr,
 	}
 }
 
@@ -614,38 +146,11 @@ func (p *plugin) GetScanJobSpec(ctx trivyoperator.PluginContext, workload client
 	if err != nil {
 		return corev1.PodSpec{}, nil, err
 	}
-
-	mode, err := config.GetMode()
-	if err != nil {
-		return corev1.PodSpec{}, nil, err
-	}
-	command, err := config.GetCommand()
-	if err != nil {
-		return corev1.PodSpec{}, nil, err
-	}
-
 	var podSpec corev1.PodSpec
 	var secrets []*corev1.Secret
-	if command == Image {
-		switch mode {
-		case Standalone:
-			podSpec, secrets, err = p.getPodSpecForStandaloneMode(ctx, config, workload, credentials, securityContext)
-		case ClientServer:
-			podSpec, secrets, err = p.getPodSpecForClientServerMode(ctx, config, workload, credentials, securityContext)
-		default:
-			return corev1.PodSpec{}, nil, fmt.Errorf("unrecognized trivy mode %q for command %q", mode, command)
-		}
-	}
-	if command == Filesystem || command == Rootfs {
-		switch mode {
-		case Standalone:
-			podSpec, secrets, err = p.getPodSpecForStandaloneFSMode(ctx, command, config, workload, securityContext)
-		case ClientServer:
-			podSpec, secrets, err = p.getPodSpecForClientServerFSMode(ctx, command, config, workload, securityContext)
-		default:
-			return corev1.PodSpec{}, nil, fmt.Errorf("unrecognized trivy mode %q for command %q", mode, command)
-		}
-	}
+
+	podSpec, secrets, err = p.podSpecMgr.GetPodSpec(ctx, config, workload, credentials, securityContext, p)
+
 	// add image pull secret to be used when pulling trivy image fom private registry
 	podSpec.ImagePullSecrets = config.GetImagePullSecret()
 	return podSpec, secrets, err
@@ -677,254 +182,6 @@ const (
 	SslCertDir                  = "/var/ssl-cert"
 )
 
-// In the Standalone mode there is the init container responsible for
-// downloading the latest Trivy DB file from GitHub and storing it to the
-// emptyDir volume shared with main containers. In other words, the init
-// container runs the following Trivy command:
-//
-//	trivy --cache-dir /tmp/trivy/.cache image --download-db-only
-//
-// The number of main containers correspond to the number of containers
-// defined for the scanned workload. Each container runs the Trivy image scan
-// command and skips the database download:
-//
-//	trivy --cache-dir /tmp/trivy/.cache image --skip-update \
-//	  --format json <container image>
-func (p *plugin) getPodSpecForStandaloneMode(ctx trivyoperator.PluginContext, config Config, workload client.Object, credentials map[string]docker.Auth, securityContext *corev1.SecurityContext) (corev1.PodSpec, []*corev1.Secret, error) {
-	var secret *corev1.Secret
-	var secrets []*corev1.Secret
-	var containersSpec []corev1.Container
-
-	spec, err := kube.GetPodSpec(workload)
-	if err != nil {
-		return corev1.PodSpec{}, nil, err
-	}
-
-	for _, c := range getContainers(spec) {
-		optionalMirroredImage, err := GetMirroredImage(c.Image, config.GetMirrors())
-		if err != nil {
-			return corev1.PodSpec{}, nil, err
-		}
-		c.Image = optionalMirroredImage
-		containersSpec = append(containersSpec, c)
-	}
-
-	containerImages := kube.GetContainerImagesFromContainersList(containersSpec)
-	containersCredentials, err := kube.MapContainerNamesToDockerAuths(containerImages, credentials)
-	if err != nil {
-		return corev1.PodSpec{}, nil, err
-	}
-	if len(containersCredentials) > 0 {
-		secret = p.newSecretWithAggregateImagePullCredentials(workload, containerImages, containersCredentials)
-		secrets = append(secrets, secret)
-	}
-
-	trivyImageRef, err := config.GetImageRef()
-	if err != nil {
-		return corev1.PodSpec{}, nil, err
-	}
-
-	trivyConfigName := trivyoperator.GetPluginConfigMapName(Plugin)
-
-	dbRepository, err := config.GetDBRepository()
-	if err != nil {
-		return corev1.PodSpec{}, nil, err
-	}
-
-	requirements, err := config.GetResourceRequirements()
-	if err != nil {
-		return corev1.PodSpec{}, nil, err
-	}
-
-	initContainer := corev1.Container{
-		Name:                     p.idGenerator.GenerateID(),
-		Image:                    trivyImageRef,
-		ImagePullPolicy:          corev1.PullPolicy(config.GetImagePullPolicy()),
-		TerminationMessagePolicy: corev1.TerminationMessageFallbackToLogsOnError,
-		Env:                      p.initContainerEnvVar(trivyConfigName, config),
-		Command: []string{
-			"trivy",
-		},
-		Args: []string{
-			"--cache-dir",
-			"/tmp/trivy/.cache",
-			"image",
-			"--download-db-only",
-			"--db-repository",
-			dbRepository,
-		},
-		Resources:       requirements,
-		SecurityContext: securityContext,
-		VolumeMounts: []corev1.VolumeMount{
-			{
-				Name:      tmpVolumeName,
-				MountPath: "/tmp",
-				ReadOnly:  false,
-			},
-		},
-	}
-
-	var containers []corev1.Container
-
-	volumeMounts := []corev1.VolumeMount{
-		{
-			Name:      tmpVolumeName,
-			ReadOnly:  false,
-			MountPath: "/tmp",
-		},
-	}
-	volumes := []corev1.Volume{
-		{
-			Name: tmpVolumeName,
-			VolumeSource: corev1.VolumeSource{
-				EmptyDir: &corev1.EmptyDirVolumeSource{
-					Medium: corev1.StorageMediumDefault,
-				},
-			},
-		},
-	}
-	volumeMounts = append(volumeMounts, getScanResultVolumeMount())
-	volumes = append(volumes, getScanResultVolume())
-
-	if volume, volumeMount := config.GenerateIgnoreFileVolumeIfAvailable(trivyConfigName); volume != nil && volumeMount != nil {
-		volumes = append(volumes, *volume)
-		volumeMounts = append(volumeMounts, *volumeMount)
-	}
-	if volume, volumeMount := config.GenerateIgnorePolicyVolumeIfAvailable(trivyConfigName, workload); volume != nil && volumeMount != nil {
-		volumes = append(volumes, *volume)
-		volumeMounts = append(volumeMounts, *volumeMount)
-	}
-	if volume, volumeMount := config.GenerateSslCertDirVolumeIfAvailable(trivyConfigName); volume != nil && volumeMount != nil {
-		volumes = append(volumes, *volume)
-		volumeMounts = append(volumeMounts, *volumeMount)
-	}
-
-	for _, c := range containersSpec {
-		env := []corev1.EnvVar{
-			constructEnvVarSourceFromConfigMap("TRIVY_SEVERITY", trivyConfigName, KeyTrivySeverity),
-			constructEnvVarSourceFromConfigMap("TRIVY_IGNORE_UNFIXED", trivyConfigName, keyTrivyIgnoreUnfixed),
-			constructEnvVarSourceFromConfigMap("TRIVY_OFFLINE_SCAN", trivyConfigName, keyTrivyOfflineScan),
-			constructEnvVarSourceFromConfigMap("TRIVY_JAVA_DB_REPOSITORY", trivyConfigName, keyTrivyJavaDBRepository),
-			constructEnvVarSourceFromConfigMap("TRIVY_TIMEOUT", trivyConfigName, keyTrivyTimeout),
-			ConfigWorkloadAnnotationEnvVars(workload, SkipFilesAnnotation, "TRIVY_SKIP_FILES", trivyConfigName, keyTrivySkipFiles),
-			ConfigWorkloadAnnotationEnvVars(workload, SkipDirsAnnotation, "TRIVY_SKIP_DIRS", trivyConfigName, keyTrivySkipDirs),
-			constructEnvVarSourceFromConfigMap("HTTP_PROXY", trivyConfigName, keyTrivyHTTPProxy),
-			constructEnvVarSourceFromConfigMap("HTTPS_PROXY", trivyConfigName, keyTrivyHTTPSProxy),
-			constructEnvVarSourceFromConfigMap("NO_PROXY", trivyConfigName, keyTrivyNoProxy),
-		}
-
-		if len(config.GetSslCertDir()) > 0 {
-			env = append(env, corev1.EnvVar{
-				Name:  "SSL_CERT_DIR",
-				Value: SslCertDir,
-			})
-		}
-		if config.IgnoreFileExists() {
-			env = append(env, corev1.EnvVar{
-				Name:  "TRIVY_IGNOREFILE",
-				Value: ignoreFileMountPath,
-			})
-		}
-		if config.FindIgnorePolicyKey(workload) != "" {
-			env = append(env, corev1.EnvVar{
-				Name:  "TRIVY_IGNORE_POLICY",
-				Value: ignorePolicyMountPath,
-			})
-		}
-
-		region := CheckAwsEcrPrivateRegistry(c.Image)
-		if region != "" {
-			env = append(env, corev1.EnvVar{
-				Name:  "AWS_REGION",
-				Value: region,
-			})
-		}
-		if config.GetDBRepositoryInsecure() {
-			env = append(env, corev1.EnvVar{
-				Name:  "TRIVY_INSECURE",
-				Value: "true",
-			})
-		}
-		gcrImage := checkGcpCrOrPivateRegistry(c.Image)
-		if _, ok := containersCredentials[c.Name]; ok && secret != nil {
-			registryUsernameKey := fmt.Sprintf("%s.username", c.Name)
-			registryPasswordKey := fmt.Sprintf("%s.password", c.Name)
-			secretName := secret.Name
-			if gcrImage {
-				createEnvandVolumeForGcr(&env, &volumeMounts, &volumes, &registryPasswordKey, &secretName)
-			} else {
-				env = append(env, corev1.EnvVar{
-					Name: "TRIVY_USERNAME",
-					ValueFrom: &corev1.EnvVarSource{
-						SecretKeyRef: &corev1.SecretKeySelector{
-							LocalObjectReference: corev1.LocalObjectReference{
-								Name: secret.Name,
-							},
-							Key: registryUsernameKey,
-						},
-					},
-				}, corev1.EnvVar{
-					Name: "TRIVY_PASSWORD",
-					ValueFrom: &corev1.EnvVarSource{
-						SecretKeyRef: &corev1.SecretKeySelector{
-							LocalObjectReference: corev1.LocalObjectReference{
-								Name: secret.Name,
-							},
-							Key: registryPasswordKey,
-						},
-					},
-				})
-			}
-
-		}
-
-		env, err = p.appendTrivyInsecureEnv(config, c.Image, env)
-		if err != nil {
-			return corev1.PodSpec{}, nil, err
-		}
-
-		env, err = p.appendTrivyNonSSLEnv(config, c.Image, env)
-		if err != nil {
-			return corev1.PodSpec{}, nil, err
-		}
-
-		resourceRequirements, err := config.GetResourceRequirements()
-		if err != nil {
-			return corev1.PodSpec{}, nil, err
-		}
-
-		imageRef, err := containerimage.ParseReference(c.Image)
-		if err != nil {
-			return corev1.PodSpec{}, nil, err
-		}
-		resultFileName := getUniqueScanResultFileName(c.Name)
-		cmd, args := p.getCommandAndArgs(ctx, Standalone, imageRef.String(), "", resultFileName)
-		containers = append(containers, corev1.Container{
-			Name:                     c.Name,
-			Image:                    trivyImageRef,
-			ImagePullPolicy:          corev1.PullPolicy(config.GetImagePullPolicy()),
-			TerminationMessagePolicy: corev1.TerminationMessageFallbackToLogsOnError,
-			Env:                      env,
-			Command:                  cmd,
-			Args:                     args,
-			Resources:                resourceRequirements,
-			SecurityContext:          securityContext,
-			VolumeMounts:             volumeMounts,
-		})
-	}
-
-	return corev1.PodSpec{
-		Affinity:                     trivyoperator.LinuxNodeAffinity(),
-		RestartPolicy:                corev1.RestartPolicyNever,
-		ServiceAccountName:           ctx.GetServiceAccountName(),
-		AutomountServiceAccountToken: pointer.Bool(getAutomountServiceAccountToken(ctx)),
-		Volumes:                      volumes,
-		InitContainers:               []corev1.Container{initContainer},
-		Containers:                   containers,
-		SecurityContext:              &corev1.PodSecurityContext{},
-	}, secrets, nil
-}
-
 func checkGcpCrOrPivateRegistry(imageUrl string) bool {
 	imageRegex := regexp.MustCompile(GCPCR_Inage_Regex)
 	return imageRegex.MatchString(imageUrl)
@@ -945,214 +202,6 @@ func (p *plugin) initContainerEnvVar(trivyConfigName string, config Config) []co
 		})
 	}
 	return envs
-}
-
-// In the ClientServer mode the number of containers of the pod created by the
-// scan job equals the number of containers defined for the scanned workload.
-// Each container runs Trivy image scan command and refers to Trivy server URL
-// returned by Config.GetServerURL:
-//
-//	trivy image --server <server URL> \
-//	  --format json <container image>
-func (p *plugin) getPodSpecForClientServerMode(ctx trivyoperator.PluginContext, config Config, workload client.Object, credentials map[string]docker.Auth, securityContext *corev1.SecurityContext) (corev1.PodSpec, []*corev1.Secret, error) {
-	var secret *corev1.Secret
-	var secrets []*corev1.Secret
-	var containersSpec []corev1.Container
-	spec, err := kube.GetPodSpec(workload)
-	if err != nil {
-		return corev1.PodSpec{}, nil, err
-	}
-
-	trivyImageRef, err := config.GetImageRef()
-	if err != nil {
-		return corev1.PodSpec{}, nil, err
-	}
-
-	trivyServerURL, err := config.GetServerURL()
-	if err != nil {
-		return corev1.PodSpec{}, nil, err
-	}
-
-	for _, c := range getContainers(spec) {
-		optionalMirroredImage, err := GetMirroredImage(c.Image, config.GetMirrors())
-		if err != nil {
-			return corev1.PodSpec{}, nil, err
-		}
-		c.Image = optionalMirroredImage
-		containersSpec = append(containersSpec, c)
-	}
-
-	containerImages := kube.GetContainerImagesFromContainersList(containersSpec)
-	containersCredentials, err := kube.MapContainerNamesToDockerAuths(containerImages, credentials)
-	if err != nil {
-		return corev1.PodSpec{}, nil, err
-	}
-	if len(containersCredentials) > 0 {
-		secret = p.newSecretWithAggregateImagePullCredentials(workload, containerImages, containersCredentials)
-		secrets = append(secrets, secret)
-	}
-
-	var containers []corev1.Container
-
-	trivyConfigName := trivyoperator.GetPluginConfigMapName(Plugin)
-	// add tmp volume mount
-	volumeMounts := []corev1.VolumeMount{
-		{
-			Name:      tmpVolumeName,
-			ReadOnly:  false,
-			MountPath: "/tmp",
-		},
-	}
-
-	// add tmp volume
-	volumes := []corev1.Volume{
-		{
-			Name: tmpVolumeName,
-			VolumeSource: corev1.VolumeSource{
-				EmptyDir: &corev1.EmptyDirVolumeSource{
-					Medium: corev1.StorageMediumDefault,
-				},
-			},
-		},
-	}
-
-	volumeMounts = append(volumeMounts, getScanResultVolumeMount())
-	volumes = append(volumes, getScanResultVolume())
-
-	if volume, volumeMount := config.GenerateIgnoreFileVolumeIfAvailable(trivyConfigName); volume != nil && volumeMount != nil {
-		volumes = append(volumes, *volume)
-		volumeMounts = append(volumeMounts, *volumeMount)
-	}
-	if volume, volumeMount := config.GenerateIgnorePolicyVolumeIfAvailable(trivyConfigName, workload); volume != nil && volumeMount != nil {
-		volumes = append(volumes, *volume)
-		volumeMounts = append(volumeMounts, *volumeMount)
-	}
-
-	if volume, volumeMount := config.GenerateSslCertDirVolumeIfAvailable(trivyConfigName); volume != nil && volumeMount != nil {
-		volumes = append(volumes, *volume)
-		volumeMounts = append(volumeMounts, *volumeMount)
-	}
-
-	for _, container := range containersSpec {
-		env := []corev1.EnvVar{
-			constructEnvVarSourceFromConfigMap("HTTP_PROXY", trivyConfigName, keyTrivyHTTPProxy),
-			constructEnvVarSourceFromConfigMap("HTTPS_PROXY", trivyConfigName, keyTrivyHTTPSProxy),
-			constructEnvVarSourceFromConfigMap("NO_PROXY", trivyConfigName, keyTrivyNoProxy),
-			constructEnvVarSourceFromConfigMap("TRIVY_SEVERITY", trivyConfigName, KeyTrivySeverity),
-			constructEnvVarSourceFromConfigMap("TRIVY_IGNORE_UNFIXED", trivyConfigName, keyTrivyIgnoreUnfixed),
-			constructEnvVarSourceFromConfigMap("TRIVY_OFFLINE_SCAN", trivyConfigName, keyTrivyOfflineScan),
-			constructEnvVarSourceFromConfigMap("TRIVY_JAVA_DB_REPOSITORY", trivyConfigName, keyTrivyJavaDBRepository),
-			constructEnvVarSourceFromConfigMap("TRIVY_TIMEOUT", trivyConfigName, keyTrivyTimeout),
-			ConfigWorkloadAnnotationEnvVars(workload, SkipFilesAnnotation, "TRIVY_SKIP_FILES", trivyConfigName, keyTrivySkipFiles),
-			ConfigWorkloadAnnotationEnvVars(workload, SkipDirsAnnotation, "TRIVY_SKIP_DIRS", trivyConfigName, keyTrivySkipDirs),
-			constructEnvVarSourceFromConfigMap("TRIVY_TOKEN_HEADER", trivyConfigName, keyTrivyServerTokenHeader),
-			constructEnvVarSourceFromSecret("TRIVY_TOKEN", trivyConfigName, keyTrivyServerToken),
-			constructEnvVarSourceFromSecret("TRIVY_CUSTOM_HEADERS", trivyConfigName, keyTrivyServerCustomHeaders),
-		}
-		if len(config.GetSslCertDir()) > 0 {
-			env = append(env, corev1.EnvVar{
-				Name:  "SSL_CERT_DIR",
-				Value: SslCertDir,
-			})
-		}
-		if config.IgnoreFileExists() {
-			env = append(env, corev1.EnvVar{
-				Name:  "TRIVY_IGNOREFILE",
-				Value: ignoreFileMountPath,
-			})
-		}
-		if config.FindIgnorePolicyKey(workload) != "" {
-			env = append(env, corev1.EnvVar{
-				Name:  "TRIVY_IGNORE_POLICY",
-				Value: ignorePolicyMountPath,
-			})
-		}
-
-		if auth, ok := containersCredentials[container.Name]; ok && secret != nil {
-			if checkGcpCrOrPivateRegistry(container.Image) && auth.Username == "_json_key" {
-				registryServiceAccountAuthKey := fmt.Sprintf("%s.password", container.Name)
-				createEnvandVolumeForGcr(&env, &volumeMounts, &volumes, &registryServiceAccountAuthKey, &secret.Name)
-			} else {
-				registryUsernameKey := fmt.Sprintf("%s.username", container.Name)
-				registryPasswordKey := fmt.Sprintf("%s.password", container.Name)
-				env = append(env, corev1.EnvVar{
-					Name: "TRIVY_USERNAME",
-					ValueFrom: &corev1.EnvVarSource{
-						SecretKeyRef: &corev1.SecretKeySelector{
-							LocalObjectReference: corev1.LocalObjectReference{
-								Name: secret.Name,
-							},
-							Key: registryUsernameKey,
-						},
-					},
-				}, corev1.EnvVar{
-					Name: "TRIVY_PASSWORD",
-					ValueFrom: &corev1.EnvVarSource{
-						SecretKeyRef: &corev1.SecretKeySelector{
-							LocalObjectReference: corev1.LocalObjectReference{
-								Name: secret.Name,
-							},
-							Key: registryPasswordKey,
-						},
-					},
-				})
-			}
-		}
-
-		env, err = p.appendTrivyInsecureEnv(config, container.Image, env)
-		if err != nil {
-			return corev1.PodSpec{}, nil, err
-		}
-
-		env, err = p.appendTrivyNonSSLEnv(config, container.Image, env)
-		if err != nil {
-			return corev1.PodSpec{}, nil, err
-		}
-
-		if config.GetServerInsecure() {
-			env = append(env, corev1.EnvVar{
-				Name:  "TRIVY_INSECURE",
-				Value: "true",
-			})
-		}
-
-		requirements, err := config.GetResourceRequirements()
-		if err != nil {
-			return corev1.PodSpec{}, nil, err
-		}
-
-		encodedTrivyServerURL, err := url.Parse(trivyServerURL)
-		if err != nil {
-			return corev1.PodSpec{}, nil, err
-		}
-		imageRef, err := containerimage.ParseReference(container.Image)
-		if err != nil {
-			return corev1.PodSpec{}, nil, err
-		}
-		resultFileName := getUniqueScanResultFileName(container.Name)
-		cmd, args := p.getCommandAndArgs(ctx, ClientServer, imageRef.String(), encodedTrivyServerURL.String(), resultFileName)
-		containers = append(containers, corev1.Container{
-			Name:                     container.Name,
-			Image:                    trivyImageRef,
-			ImagePullPolicy:          corev1.PullPolicy(config.GetImagePullPolicy()),
-			TerminationMessagePolicy: corev1.TerminationMessageFallbackToLogsOnError,
-			Env:                      env,
-			Command:                  cmd,
-			Args:                     args,
-			Resources:                requirements,
-			SecurityContext:          securityContext,
-			VolumeMounts:             volumeMounts,
-		})
-	}
-
-	return corev1.PodSpec{
-		Affinity:                     trivyoperator.LinuxNodeAffinity(),
-		RestartPolicy:                corev1.RestartPolicyNever,
-		ServiceAccountName:           ctx.GetServiceAccountName(),
-		AutomountServiceAccountToken: pointer.Bool(getAutomountServiceAccountToken(ctx)),
-		Containers:                   containers,
-		Volumes:                      volumes,
-	}, secrets, nil
 }
 
 func (p *plugin) getCommandAndArgs(ctx trivyoperator.PluginContext, mode Mode, imageRef string, trivyServerURL string, resultFileName string) ([]string, []string) {
@@ -1334,430 +383,6 @@ func createEnvandVolumeForGcr(env *[]corev1.EnvVar, volumeMounts *[]corev1.Volum
 	}
 	*volumes = append(*volumes, googlecredVolume)
 	*volumeMounts = append(*volumeMounts, googlecredMount)
-}
-
-// FileSystem scan option with standalone mode.
-// The only difference is that instead of scanning the resource by name,
-// We are scanning the resource place on a specific file system location using the following command.
-//
-//	trivy --quiet fs  --format json --ignore-unfixed  file/system/location
-func (p *plugin) getPodSpecForStandaloneFSMode(ctx trivyoperator.PluginContext, command Command, config Config,
-	workload client.Object, securityContext *corev1.SecurityContext) (corev1.PodSpec, []*corev1.Secret, error) {
-	var secrets []*corev1.Secret
-	spec, err := kube.GetPodSpec(workload)
-	if err != nil {
-		return corev1.PodSpec{}, nil, err
-	}
-	pullPolicy := corev1.PullIfNotPresent
-	// nodeName to schedule scan job explicitly on specific node.
-	var nodeName string
-	if !ctx.GetTrivyOperatorConfig().VulnerabilityScanJobsInSameNamespace() {
-		// get nodeName from running pods.
-		nodeName, err = p.objectResolver.GetNodeName(context.Background(), workload)
-		if err != nil {
-			return corev1.PodSpec{}, nil, fmt.Errorf("failed resolving node name for workload %q: %w",
-				workload.GetNamespace()+"/"+workload.GetName(), err)
-		}
-		pullPolicy = corev1.PullNever
-	}
-
-	trivyImageRef, err := config.GetImageRef()
-	if err != nil {
-		return corev1.PodSpec{}, nil, err
-	}
-
-	trivyConfigName := trivyoperator.GetPluginConfigMapName(Plugin)
-
-	dbRepository, err := config.GetDBRepository()
-	if err != nil {
-		return corev1.PodSpec{}, nil, err
-	}
-
-	requirements, err := config.GetResourceRequirements()
-	if err != nil {
-		return corev1.PodSpec{}, nil, err
-	}
-
-	volumeMounts := []corev1.VolumeMount{
-		{
-			Name:      FsSharedVolumeName,
-			ReadOnly:  false,
-			MountPath: "/var/trivyoperator",
-		},
-		{
-			Name:      tmpVolumeName,
-			MountPath: "/tmp",
-			ReadOnly:  false,
-		},
-	}
-
-	initContainerCopyBinary := corev1.Container{
-		Name:                     p.idGenerator.GenerateID(),
-		Image:                    trivyImageRef,
-		ImagePullPolicy:          corev1.PullPolicy(config.GetImagePullPolicy()),
-		TerminationMessagePolicy: corev1.TerminationMessageFallbackToLogsOnError,
-		Command: []string{
-			"cp",
-			"-v",
-			"/usr/local/bin/trivy",
-			SharedVolumeLocationOfTrivy,
-		},
-		Resources:       requirements,
-		SecurityContext: securityContext,
-		VolumeMounts:    volumeMounts,
-	}
-
-	initContainerDB := corev1.Container{
-		Name:                     p.idGenerator.GenerateID(),
-		Image:                    trivyImageRef,
-		ImagePullPolicy:          corev1.PullPolicy(config.GetImagePullPolicy()),
-		TerminationMessagePolicy: corev1.TerminationMessageFallbackToLogsOnError,
-		Env:                      p.initContainerFSEnvVar(trivyConfigName, config),
-		Command: []string{
-			"trivy",
-		},
-		Args: []string{
-			"--cache-dir",
-			"/var/trivyoperator/trivy-db",
-			"image",
-			"--download-db-only",
-			"--db-repository",
-			dbRepository,
-		},
-		Resources:       requirements,
-		SecurityContext: securityContext,
-		VolumeMounts:    volumeMounts,
-	}
-
-	var containers []corev1.Container
-
-	volumes := []corev1.Volume{
-		{
-			Name: FsSharedVolumeName,
-			VolumeSource: corev1.VolumeSource{
-				EmptyDir: &corev1.EmptyDirVolumeSource{
-					Medium: corev1.StorageMediumDefault,
-				},
-			},
-		},
-		{
-			Name: tmpVolumeName,
-			VolumeSource: corev1.VolumeSource{
-				EmptyDir: &corev1.EmptyDirVolumeSource{
-					Medium: corev1.StorageMediumDefault,
-				},
-			},
-		},
-	}
-
-	volumeMounts = append(volumeMounts, getScanResultVolumeMount())
-	volumes = append(volumes, getScanResultVolume())
-
-	if volume, volumeMount := config.GenerateIgnoreFileVolumeIfAvailable(trivyConfigName); volume != nil && volumeMount != nil {
-		volumes = append(volumes, *volume)
-		volumeMounts = append(volumeMounts, *volumeMount)
-	}
-	if volume, volumeMount := config.GenerateIgnorePolicyVolumeIfAvailable(trivyConfigName, workload); volume != nil && volumeMount != nil {
-		volumes = append(volumes, *volume)
-		volumeMounts = append(volumeMounts, *volumeMount)
-	}
-	if volume, volumeMount := config.GenerateSslCertDirVolumeIfAvailable(trivyConfigName); volume != nil && volumeMount != nil {
-		volumes = append(volumes, *volume)
-		volumeMounts = append(volumeMounts, *volumeMount)
-	}
-
-	for _, c := range getContainers(spec) {
-		env := []corev1.EnvVar{
-			constructEnvVarSourceFromConfigMap("TRIVY_SEVERITY", trivyConfigName, KeyTrivySeverity),
-			ConfigWorkloadAnnotationEnvVars(workload, SkipFilesAnnotation, "TRIVY_SKIP_FILES", trivyConfigName, keyTrivySkipFiles),
-			ConfigWorkloadAnnotationEnvVars(workload, SkipDirsAnnotation, "TRIVY_SKIP_DIRS", trivyConfigName, keyTrivySkipDirs),
-			constructEnvVarSourceFromConfigMap("HTTP_PROXY", trivyConfigName, keyTrivyHTTPProxy),
-			constructEnvVarSourceFromConfigMap("TRIVY_TIMEOUT", trivyConfigName, keyTrivyTimeout),
-			constructEnvVarSourceFromConfigMap("HTTPS_PROXY", trivyConfigName, keyTrivyHTTPSProxy),
-			constructEnvVarSourceFromConfigMap("NO_PROXY", trivyConfigName, keyTrivyNoProxy),
-			constructEnvVarSourceFromConfigMap("TRIVY_JAVA_DB_REPOSITORY", trivyConfigName, keyTrivyJavaDBRepository),
-		}
-		if len(config.GetSslCertDir()) > 0 {
-			env = append(env, corev1.EnvVar{
-				Name:  "SSL_CERT_DIR",
-				Value: SslCertDir,
-			})
-		}
-		if config.IgnoreFileExists() {
-			env = append(env, corev1.EnvVar{
-				Name:  "TRIVY_IGNOREFILE",
-				Value: ignoreFileMountPath,
-			})
-		}
-		if config.FindIgnorePolicyKey(workload) != "" {
-			env = append(env, corev1.EnvVar{
-				Name:  "TRIVY_IGNORE_POLICY",
-				Value: ignorePolicyMountPath,
-			})
-		}
-		if config.IgnoreUnfixed() {
-			env = append(env, constructEnvVarSourceFromConfigMap("TRIVY_IGNORE_UNFIXED",
-				trivyConfigName, keyTrivyIgnoreUnfixed))
-		}
-		if config.GetDBRepositoryInsecure() {
-			env = append(env, corev1.EnvVar{
-				Name:  "TRIVY_INSECURE",
-				Value: "true",
-			})
-		}
-
-		if config.OfflineScan() {
-			env = append(env, constructEnvVarSourceFromConfigMap("TRIVY_OFFLINE_SCAN",
-				trivyConfigName, keyTrivyOfflineScan))
-		}
-
-		env, err = p.appendTrivyInsecureEnv(config, c.Image, env)
-		if err != nil {
-			return corev1.PodSpec{}, nil, err
-		}
-
-		resourceRequirements, err := config.GetResourceRequirements()
-		if err != nil {
-			return corev1.PodSpec{}, nil, err
-		}
-		containers = append(containers, corev1.Container{
-			Name:                     c.Name,
-			Image:                    c.Image,
-			ImagePullPolicy:          pullPolicy,
-			TerminationMessagePolicy: corev1.TerminationMessageFallbackToLogsOnError,
-			Env:                      env,
-			Command: []string{
-				SharedVolumeLocationOfTrivy,
-			},
-			Args:            p.getFSScanningArgs(ctx, command, Standalone, ""),
-			Resources:       resourceRequirements,
-			SecurityContext: securityContext,
-			VolumeMounts:    volumeMounts,
-		})
-	}
-
-	podSpec := corev1.PodSpec{
-		Affinity:                     trivyoperator.LinuxNodeAffinity(),
-		RestartPolicy:                corev1.RestartPolicyNever,
-		ServiceAccountName:           ctx.GetServiceAccountName(),
-		AutomountServiceAccountToken: pointer.Bool(getAutomountServiceAccountToken(ctx)),
-		Volumes:                      volumes,
-		InitContainers:               []corev1.Container{initContainerCopyBinary, initContainerDB},
-		Containers:                   containers,
-		SecurityContext:              &corev1.PodSecurityContext{},
-	}
-
-	if !ctx.GetTrivyOperatorConfig().VulnerabilityScanJobsInSameNamespace() {
-		// schedule scan job explicitly on specific node.
-		podSpec.NodeName = nodeName
-	}
-
-	return podSpec, secrets, nil
-}
-
-// FileSystem scan option with ClientServer mode.
-// The only difference is that instead of scanning the resource by name,
-// We scanning the resource place on a specific file system location using the following command.
-//
-//	trivy --quiet fs  --server TRIVY_SERVER  --format json --ignore-unfixed  file/system/location
-func (p *plugin) getPodSpecForClientServerFSMode(ctx trivyoperator.PluginContext, command Command, config Config,
-	workload client.Object, securityContext *corev1.SecurityContext) (corev1.PodSpec, []*corev1.Secret, error) {
-	var secrets []*corev1.Secret
-	spec, err := kube.GetPodSpec(workload)
-	if err != nil {
-		return corev1.PodSpec{}, nil, err
-	}
-	pullPolicy := corev1.PullIfNotPresent
-	// nodeName to schedule scan job explicitly on specific node.
-	var nodeName string
-	if !ctx.GetTrivyOperatorConfig().VulnerabilityScanJobsInSameNamespace() {
-		// get nodeName from running pods.
-		nodeName, err = p.objectResolver.GetNodeName(context.Background(), workload)
-		if err != nil {
-			return corev1.PodSpec{}, nil, fmt.Errorf("failed resolving node name for workload %q: %w",
-				workload.GetNamespace()+"/"+workload.GetName(), err)
-		}
-		pullPolicy = corev1.PullNever
-	}
-
-	trivyImageRef, err := config.GetImageRef()
-	if err != nil {
-		return corev1.PodSpec{}, nil, err
-	}
-
-	trivyServerURL, err := config.GetServerURL()
-	if err != nil {
-		return corev1.PodSpec{}, nil, err
-	}
-
-	encodedTrivyServerURL, err := url.Parse(trivyServerURL)
-	if err != nil {
-		return corev1.PodSpec{}, nil, err
-	}
-
-	trivyConfigName := trivyoperator.GetPluginConfigMapName(Plugin)
-
-	requirements, err := config.GetResourceRequirements()
-	if err != nil {
-		return corev1.PodSpec{}, nil, err
-	}
-
-	volumeMounts := []corev1.VolumeMount{
-		{
-			Name:      FsSharedVolumeName,
-			ReadOnly:  false,
-			MountPath: "/var/trivyoperator",
-		},
-		{
-			Name:      tmpVolumeName,
-			MountPath: "/tmp",
-			ReadOnly:  false,
-		},
-	}
-
-	initContainerCopyBinary := corev1.Container{
-		Name:                     p.idGenerator.GenerateID(),
-		Image:                    trivyImageRef,
-		ImagePullPolicy:          corev1.PullPolicy(config.GetImagePullPolicy()),
-		TerminationMessagePolicy: corev1.TerminationMessageFallbackToLogsOnError,
-		Command: []string{
-			"cp",
-			"-v",
-			"/usr/local/bin/trivy",
-			SharedVolumeLocationOfTrivy,
-		},
-		Resources:       requirements,
-		SecurityContext: securityContext,
-		VolumeMounts:    volumeMounts,
-	}
-
-	var containers []corev1.Container
-
-	volumes := []corev1.Volume{
-		{
-			Name: FsSharedVolumeName,
-			VolumeSource: corev1.VolumeSource{
-				EmptyDir: &corev1.EmptyDirVolumeSource{
-					Medium: corev1.StorageMediumDefault,
-				},
-			},
-		},
-		{
-			Name: tmpVolumeName,
-			VolumeSource: corev1.VolumeSource{
-				EmptyDir: &corev1.EmptyDirVolumeSource{
-					Medium: corev1.StorageMediumDefault,
-				},
-			},
-		},
-	}
-	volumeMounts = append(volumeMounts, getScanResultVolumeMount())
-	volumes = append(volumes, getScanResultVolume())
-
-	if volume, volumeMount := config.GenerateIgnoreFileVolumeIfAvailable(trivyConfigName); volume != nil && volumeMount != nil {
-		volumes = append(volumes, *volume)
-		volumeMounts = append(volumeMounts, *volumeMount)
-	}
-	if volume, volumeMount := config.GenerateIgnorePolicyVolumeIfAvailable(trivyConfigName, workload); volume != nil && volumeMount != nil {
-		volumes = append(volumes, *volume)
-		volumeMounts = append(volumeMounts, *volumeMount)
-	}
-	if volume, volumeMount := config.GenerateSslCertDirVolumeIfAvailable(trivyConfigName); volume != nil && volumeMount != nil {
-		volumes = append(volumes, *volume)
-		volumeMounts = append(volumeMounts, *volumeMount)
-	}
-
-	for _, c := range getContainers(spec) {
-		env := []corev1.EnvVar{
-			constructEnvVarSourceFromConfigMap("TRIVY_SEVERITY", trivyConfigName, KeyTrivySeverity),
-			ConfigWorkloadAnnotationEnvVars(workload, SkipFilesAnnotation, "TRIVY_SKIP_FILES", trivyConfigName, keyTrivySkipFiles),
-			ConfigWorkloadAnnotationEnvVars(workload, SkipDirsAnnotation, "TRIVY_SKIP_DIRS", trivyConfigName, keyTrivySkipDirs),
-			constructEnvVarSourceFromConfigMap("HTTP_PROXY", trivyConfigName, keyTrivyHTTPProxy),
-			constructEnvVarSourceFromConfigMap("TRIVY_TIMEOUT", trivyConfigName, keyTrivyTimeout),
-			constructEnvVarSourceFromConfigMap("HTTPS_PROXY", trivyConfigName, keyTrivyHTTPSProxy),
-			constructEnvVarSourceFromConfigMap("NO_PROXY", trivyConfigName, keyTrivyNoProxy),
-			constructEnvVarSourceFromConfigMap("TRIVY_TOKEN_HEADER", trivyConfigName, keyTrivyServerTokenHeader),
-			constructEnvVarSourceFromSecret("TRIVY_TOKEN", trivyConfigName, keyTrivyServerToken),
-			constructEnvVarSourceFromSecret("TRIVY_CUSTOM_HEADERS", trivyConfigName, keyTrivyServerCustomHeaders),
-			constructEnvVarSourceFromConfigMap("TRIVY_JAVA_DB_REPOSITORY", trivyConfigName, keyTrivyJavaDBRepository),
-		}
-		if len(config.GetSslCertDir()) > 0 {
-			env = append(env, corev1.EnvVar{
-				Name:  "SSL_CERT_DIR",
-				Value: SslCertDir,
-			})
-		}
-		if config.IgnoreFileExists() {
-			env = append(env, corev1.EnvVar{
-				Name:  "TRIVY_IGNOREFILE",
-				Value: ignoreFileMountPath,
-			})
-		}
-		if config.FindIgnorePolicyKey(workload) != "" {
-			env = append(env, corev1.EnvVar{
-				Name:  "TRIVY_IGNORE_POLICY",
-				Value: ignorePolicyMountPath,
-			})
-		}
-		if config.IgnoreUnfixed() {
-			env = append(env, constructEnvVarSourceFromConfigMap("TRIVY_IGNORE_UNFIXED",
-				trivyConfigName, keyTrivyIgnoreUnfixed))
-		}
-
-		if config.OfflineScan() {
-			env = append(env, constructEnvVarSourceFromConfigMap("TRIVY_OFFLINE_SCAN",
-				trivyConfigName, keyTrivyOfflineScan))
-		}
-
-		env, err = p.appendTrivyInsecureEnv(config, c.Image, env)
-		if err != nil {
-			return corev1.PodSpec{}, nil, err
-		}
-
-		if config.GetServerInsecure() {
-			env = append(env, corev1.EnvVar{
-				Name:  "TRIVY_INSECURE",
-				Value: "true",
-			})
-		}
-
-		resourceRequirements, err := config.GetResourceRequirements()
-		if err != nil {
-			return corev1.PodSpec{}, nil, err
-		}
-		containers = append(containers, corev1.Container{
-			Name:                     c.Name,
-			Image:                    c.Image,
-			ImagePullPolicy:          pullPolicy,
-			TerminationMessagePolicy: corev1.TerminationMessageFallbackToLogsOnError,
-			Env:                      env,
-			Command: []string{
-				SharedVolumeLocationOfTrivy,
-			},
-			Args:            p.getFSScanningArgs(ctx, command, ClientServer, encodedTrivyServerURL.String()),
-			Resources:       resourceRequirements,
-			SecurityContext: securityContext,
-			VolumeMounts:    volumeMounts,
-		})
-	}
-
-	podSpec := corev1.PodSpec{
-		Affinity:                     trivyoperator.LinuxNodeAffinity(),
-		RestartPolicy:                corev1.RestartPolicyNever,
-		ServiceAccountName:           ctx.GetServiceAccountName(),
-		AutomountServiceAccountToken: pointer.Bool(getAutomountServiceAccountToken(ctx)),
-		Volumes:                      volumes,
-		InitContainers:               []corev1.Container{initContainerCopyBinary},
-		Containers:                   containers,
-		SecurityContext:              &corev1.PodSecurityContext{},
-	}
-
-	if !ctx.GetTrivyOperatorConfig().VulnerabilityScanJobsInSameNamespace() {
-		// schedule scan job explicitly on specific node.
-		podSpec.NodeName = nodeName
-	}
-
-	return podSpec, secrets, nil
 }
 
 func (p *plugin) getFSScanningArgs(ctx trivyoperator.PluginContext, command Command, mode Mode, trivyServerURL string) []string {
