+++ pkg/plugins/trivy/plugin_test.go
@@ -2,7 +2,7 @@ package trivy_test
 
 import (
 	"bytes"
-	"context"
+
 	"encoding/base64"
 	"errors"
 	"fmt"
@@ -32,7 +32,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
+
 	"k8s.io/utils/pointer"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
@@ -42,790 +42,6 @@ var (
 	fixedTime  = time.Now()
 	fixedClock = ext.NewFixedClock(fixedTime)
 )
-
-func TestConfig_GetImageRef(t *testing.T) {
-	testCases := []struct {
-		name             string
-		configData       trivy.Config
-		expectedError    string
-		expectedImageRef string
-	}{
-		{
-			name:          "Should return error",
-			configData:    trivy.Config{PluginConfig: trivyoperator.PluginConfig{}},
-			expectedError: "property trivy.repository not set",
-		},
-		{
-			name: "Should return error",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.tag": "0.8.0",
-				},
-			}},
-			expectedError: "property trivy.repository not set",
-		},
-		{
-			name: "Should return error",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.repository": "gcr.io/aquasecurity/trivy",
-				},
-			}},
-			expectedError: "property trivy.tag not set",
-		},
-		{
-			name: "Should return image reference from config data",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.repository": "gcr.io/aquasecurity/trivy",
-					"trivy.tag":        "0.8.0",
-				},
-			}},
-			expectedImageRef: "gcr.io/aquasecurity/trivy:0.8.0",
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			imageRef, err := tc.configData.GetImageRef()
-			if tc.expectedError != "" {
-				require.EqualError(t, err, tc.expectedError)
-			} else {
-				require.NoError(t, err)
-				assert.Equal(t, tc.expectedImageRef, imageRef)
-			}
-		})
-	}
-}
-
-func TestConfig_GetAdditionalVulnerabilityReportFields(t *testing.T) {
-	testCases := []struct {
-		name             string
-		configData       trivy.Config
-		additionalFields trivy.AdditionalFields
-	}{
-		{
-			name:             "no additional fields are set",
-			configData:       trivy.Config{PluginConfig: trivyoperator.PluginConfig{}},
-			additionalFields: trivy.AdditionalFields{},
-		},
-		{
-			name: "all additional fields are set",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.additionalVulnerabilityReportFields": "PackageType,PkgPath,Class,Target,Links,Description,CVSS",
-				},
-			}},
-			additionalFields: trivy.AdditionalFields{Description: true, Links: true, CVSS: true, Class: true, PackageType: true, PkgPath: true, Target: true},
-		},
-		{
-			name: "some additional fields are set",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.additionalVulnerabilityReportFields": "PackageType,Target,Links,CVSS",
-				},
-			}},
-			additionalFields: trivy.AdditionalFields{Links: true, CVSS: true, PackageType: true, Target: true},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			addFields := tc.configData.GetAdditionalVulnerabilityReportFields()
-			assert.True(t, addFields.Description == tc.additionalFields.Description)
-			assert.True(t, addFields.CVSS == tc.additionalFields.CVSS)
-			assert.True(t, addFields.Target == tc.additionalFields.Target)
-			assert.True(t, addFields.PackageType == tc.additionalFields.PackageType)
-			assert.True(t, addFields.Class == tc.additionalFields.Class)
-			assert.True(t, addFields.Links == tc.additionalFields.Links)
-		})
-	}
-}
-
-func TestConfig_GetMode(t *testing.T) {
-	testCases := []struct {
-		name          string
-		configData    trivy.Config
-		expectedError string
-		expectedMode  trivy.Mode
-	}{
-		{
-			name: "Should return Standalone",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.mode": string(trivy.Standalone),
-				},
-			}},
-			expectedMode: trivy.Standalone,
-		},
-		{
-			name: "Should return ClientServer",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.mode": string(trivy.ClientServer),
-				},
-			}},
-			expectedMode: trivy.ClientServer,
-		},
-		{
-			name:          "Should return error when value is not set",
-			configData:    trivy.Config{PluginConfig: trivyoperator.PluginConfig{}},
-			expectedError: "property trivy.mode not set",
-		},
-		{
-			name: "Should return error when value is not allowed",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.mode": "P2P",
-				},
-			}},
-			expectedError: "invalid value (P2P) of trivy.mode; allowed values (Standalone, ClientServer)",
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			mode, err := tc.configData.GetMode()
-			if tc.expectedError != "" {
-				require.EqualError(t, err, tc.expectedError)
-			} else {
-				require.NoError(t, err)
-				assert.Equal(t, tc.expectedMode, mode)
-			}
-		})
-	}
-}
-
-func TestGetSlow(t *testing.T) {
-	testCases := []struct {
-		name       string
-		configData trivy.Config
-		want       bool
-	}{
-		{
-			name: "slow param set to true",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.slow": "true",
-				},
-			}},
-			want: true,
-		},
-		{
-			name: "slow param set to false",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.slow": "false",
-				},
-			}},
-			want: false,
-		},
-		{
-			name: "slow param set to no valid value",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.slow": "false2",
-				},
-			}},
-			want: true,
-		},
-		{
-			name: "slow param set to no  value",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{},
-			}},
-			want: true,
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			got := tc.configData.GetSlow()
-			assert.Equal(t, got, tc.want)
-
-		})
-	}
-}
-func TestConfig_GetCommand(t *testing.T) {
-	testCases := []struct {
-		name            string
-		configData      trivy.Config
-		expectedError   string
-		expectedCommand trivy.Command
-	}{
-		{
-			name: "Should return image",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.command": "image",
-				},
-			}},
-			expectedCommand: trivy.Image,
-		},
-		{
-			name: "Should return image when value is not set",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{},
-			}},
-			expectedCommand: trivy.Image,
-		},
-		{
-			name: "Should return filesystem",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.command": "filesystem",
-				},
-			}},
-			expectedCommand: trivy.Filesystem,
-		},
-		{
-			name: "Should return rootfs",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.command": "rootfs",
-				},
-			}},
-			expectedCommand: trivy.Rootfs,
-		},
-		{
-			name: "Should return error when value is not allowed",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.command": "ls",
-				},
-			}},
-			expectedError: "invalid value (ls) of trivy.command; allowed values (image, filesystem, rootfs)",
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			command, err := tc.configData.GetCommand()
-			if tc.expectedError != "" {
-				require.EqualError(t, err, tc.expectedError)
-			} else {
-				require.NoError(t, err)
-				assert.Equal(t, tc.expectedCommand, command)
-			}
-		})
-	}
-}
-
-func TestVulnType(t *testing.T) {
-	testCases := []struct {
-		name       string
-		configData trivy.Config
-		want       string
-	}{
-		{
-			name: "valid vuln type os",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.vulnType": "os",
-				},
-			}},
-			want: "os",
-		},
-		{
-			name: "valid vuln type library",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.vulnType": "library",
-				},
-			}},
-			want: "library",
-		},
-		{
-			name: "empty vuln type",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.vulnType": "",
-				},
-			}},
-			want: "",
-		},
-		{
-			name: "non valid vuln type",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.vulnType": "aaa",
-				},
-			}},
-			want: "",
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			got := tc.configData.GetVulnType()
-			assert.Equal(t, got, tc.want)
-
-		})
-	}
-}
-
-func TestConfig_GetResourceRequirements(t *testing.T) {
-	testCases := []struct {
-		name                 string
-		config               trivy.Config
-		expectedError        string
-		expectedRequirements corev1.ResourceRequirements
-	}{
-		{
-			name: "Should return empty requirements by default",
-			config: trivy.Config{
-				PluginConfig: trivyoperator.PluginConfig{},
-			},
-			expectedError: "",
-			expectedRequirements: corev1.ResourceRequirements{
-				Requests: corev1.ResourceList{},
-				Limits:   corev1.ResourceList{},
-			},
-		},
-		{
-			name: "Should return configured resource requirement",
-			config: trivy.Config{
-				PluginConfig: trivyoperator.PluginConfig{
-					Data: map[string]string{
-						"trivy.dbRepository":              trivy.DefaultDBRepository,
-						"trivy.javaDbRepository":          trivy.DefaultJavaDBRepository,
-						"trivy.resources.requests.cpu":    "800m",
-						"trivy.resources.requests.memory": "200M",
-						"trivy.resources.limits.cpu":      "600m",
-						"trivy.resources.limits.memory":   "700M",
-					},
-				},
-			},
-			expectedError: "",
-			expectedRequirements: corev1.ResourceRequirements{
-				Requests: corev1.ResourceList{
-					corev1.ResourceCPU:    resource.MustParse("800m"),
-					corev1.ResourceMemory: resource.MustParse("200M"),
-				},
-				Limits: corev1.ResourceList{
-					corev1.ResourceCPU:    resource.MustParse("600m"),
-					corev1.ResourceMemory: resource.MustParse("700M"),
-				},
-			},
-		},
-		{
-			name: "Should return error if resource is not parseable",
-			config: trivy.Config{
-				PluginConfig: trivyoperator.PluginConfig{
-					Data: map[string]string{
-						"trivy.resources.requests.cpu": "roughly 100",
-					},
-				},
-			},
-			expectedError: "parsing resource definition trivy.resources.requests.cpu: roughly 100 quantities must match the regular expression '^([+-]?[0-9.]+)([eEinumkKMGTP]*[-+]?[0-9]*)$'",
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			resourceRequirement, err := tc.config.GetResourceRequirements()
-			if tc.expectedError != "" {
-				require.EqualError(t, err, tc.expectedError)
-			} else {
-				require.NoError(t, err)
-				assert.Equal(t, tc.expectedRequirements, resourceRequirement, tc.name)
-			}
-		})
-	}
-}
-
-func TestConfig_IgnoreFileExists(t *testing.T) {
-	testCases := []struct {
-		name           string
-		configData     trivy.Config
-		expectedOutput bool
-	}{
-		{
-			name: "Should return false",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"foo": "bar",
-				},
-			}},
-			expectedOutput: false,
-		},
-		{
-			name: "Should return true",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"foo": "bar",
-					"trivy.ignoreFile": `# Accept the risk
-CVE-2018-14618
-
-# No impact in our settings
-CVE-2019-1543`,
-				},
-			}},
-			expectedOutput: true,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			exists := tc.configData.IgnoreFileExists()
-			assert.Equal(t, tc.expectedOutput, exists)
-		})
-	}
-}
-
-func TestConfig_IgnoreUnfixed(t *testing.T) {
-	testCases := []struct {
-		name           string
-		configData     trivy.Config
-		expectedOutput bool
-	}{
-		{
-			name: "Should return false",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"foo": "bar",
-				},
-			}},
-			expectedOutput: false,
-		},
-		{
-			name: "Should return true",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"foo":                 "bar",
-					"trivy.ignoreUnfixed": "true",
-				},
-			}},
-			expectedOutput: true,
-		},
-		{
-			name: "Should return false when set it as false",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"foo":                 "bar",
-					"trivy.ignoreUnfixed": "false",
-				},
-			}},
-			expectedOutput: true,
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			exists := tc.configData.IgnoreUnfixed()
-			assert.Equal(t, tc.expectedOutput, exists)
-		})
-	}
-}
-
-func TestConfig_OfflineScan(t *testing.T) {
-	testCases := []struct {
-		name           string
-		configData     trivy.Config
-		expectedOutput bool
-	}{
-		{
-			name: "Should return false",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"foo": "bar",
-				},
-			}},
-			expectedOutput: false,
-		},
-		{
-			name: "Should return true",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"foo":               "bar",
-					"trivy.offlineScan": "true",
-				},
-			}},
-			expectedOutput: true,
-		},
-		{
-			name: "Should return false when set it as false",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"foo":               "bar",
-					"trivy.offlineScan": "false",
-				},
-			}},
-			expectedOutput: true,
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			exists := tc.configData.OfflineScan()
-			assert.Equal(t, tc.expectedOutput, exists)
-		})
-	}
-}
-
-func TestConfig_dbRepositoryInsecure(t *testing.T) {
-	testCases := []struct {
-		name           string
-		configData     trivy.Config
-		expectedOutput bool
-	}{
-		{
-			name: "good value Should return false",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.dbRepositoryInsecure": "false",
-				},
-			}},
-			expectedOutput: false,
-		},
-		{
-			name: "good value Should return true",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.dbRepositoryInsecure": "true",
-				},
-			}},
-			expectedOutput: true,
-		},
-		{
-			name: "bad value Should return false",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.dbRepositoryInsecure": "true1",
-				},
-			}},
-			expectedOutput: false,
-		},
-		{
-			name: "no value Should return false",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{},
-			}},
-			expectedOutput: false,
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			exists := tc.configData.GetDBRepositoryInsecure()
-			assert.Equal(t, tc.expectedOutput, exists)
-		})
-	}
-}
-
-func TestConfig_GetInsecureRegistries(t *testing.T) {
-	testCases := []struct {
-		name           string
-		configData     trivy.Config
-		expectedOutput map[string]bool
-	}{
-		{
-			name: "Should return nil map when there is no key with trivy.insecureRegistry. prefix",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"foo": "bar",
-				},
-			}},
-			expectedOutput: make(map[string]bool),
-		},
-		{
-			name: "Should return insecure registries in map",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"foo":                                "bar",
-					"trivy.insecureRegistry.pocRegistry": "poc.myregistry.harbor.com.pl",
-					"trivy.insecureRegistry.qaRegistry":  "qa.registry.aquasec.com",
-				},
-			}},
-			expectedOutput: map[string]bool{
-				"poc.myregistry.harbor.com.pl": true,
-				"qa.registry.aquasec.com":      true,
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			insecureRegistries := tc.configData.GetInsecureRegistries()
-			assert.Equal(t, tc.expectedOutput, insecureRegistries)
-		})
-	}
-}
-
-func TestConfig_GetNonSSLRegistries(t *testing.T) {
-	testCases := []struct {
-		name           string
-		configData     trivy.Config
-		expectedOutput map[string]bool
-	}{
-		{
-			name: "Should return nil map when there is no key with trivy.nonSslRegistry. prefix",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"foo": "bar",
-				},
-			}},
-			expectedOutput: make(map[string]bool),
-		},
-		{
-			name: "Should return insecure registries in map",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"foo":                              "bar",
-					"trivy.nonSslRegistry.pocRegistry": "poc.myregistry.harbor.com.pl",
-					"trivy.nonSslRegistry.qaRegistry":  "qa.registry.aquasec.com",
-				},
-			}},
-			expectedOutput: map[string]bool{
-				"poc.myregistry.harbor.com.pl": true,
-				"qa.registry.aquasec.com":      true,
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			nonSslRegistries := tc.configData.GetNonSSLRegistries()
-			assert.Equal(t, tc.expectedOutput, nonSslRegistries)
-		})
-	}
-}
-
-func TestConfig_GetMirrors(t *testing.T) {
-	testCases := []struct {
-		name           string
-		configData     trivy.Config
-		expectedOutput map[string]string
-	}{
-		{
-			name: "Should return empty map when there is no key with trivy.mirrors.registry. prefix",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"foo": "bar",
-				},
-			}},
-			expectedOutput: make(map[string]string),
-		},
-		{
-			name: "Should return mirrors in a map",
-			configData: trivy.Config{PluginConfig: trivyoperator.PluginConfig{
-				Data: map[string]string{
-					"trivy.registry.mirror.docker.io": "mirror.io",
-				},
-			}},
-			expectedOutput: map[string]string{"docker.io": "mirror.io"},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			assert.Equal(t, tc.expectedOutput, tc.configData.GetMirrors())
-		})
-	}
-}
-
-func TestPlugin_Init(t *testing.T) {
-
-	t.Run("Should create the default config", func(t *testing.T) {
-		testClient := fake.NewClientBuilder().WithObjects().Build()
-		or := kube.NewObjectResolver(testClient, &kube.CompatibleObjectMapper{})
-		instance := trivy.NewPlugin(fixedClock, ext.NewSimpleIDGenerator(), &or)
-
-		pluginContext := trivyoperator.NewPluginContext().
-			WithName(trivy.Plugin).
-			WithNamespace("trivyoperator-ns").
-			WithServiceAccountName("trivyoperator-sa").
-			WithClient(testClient).
-			Get()
-		err := instance.Init(pluginContext)
-		require.NoError(t, err)
-
-		var cm corev1.ConfigMap
-		err = testClient.Get(context.Background(), types.NamespacedName{
-			Namespace: "trivyoperator-ns",
-			Name:      "trivy-operator-trivy-config",
-		}, &cm)
-		require.NoError(t, err)
-		assert.Equal(t, corev1.ConfigMap{
-			TypeMeta: metav1.TypeMeta{
-				APIVersion: "v1",
-				Kind:       "ConfigMap",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "trivy-operator-trivy-config",
-				Namespace: "trivyoperator-ns",
-				Labels: map[string]string{
-					"app.kubernetes.io/managed-by": "trivyoperator",
-				},
-				ResourceVersion: "1",
-			},
-			Data: map[string]string{
-				"trivy.repository":                trivy.DefaultImageRepository,
-				"trivy.tag":                       "0.45.1",
-				"trivy.severity":                  trivy.DefaultSeverity,
-				"trivy.slow":                      "true",
-				"trivy.mode":                      string(trivy.Standalone),
-				"trivy.timeout":                   "5m0s",
-				"trivy.dbRepository":              trivy.DefaultDBRepository,
-				"trivy.javaDbRepository":          trivy.DefaultJavaDBRepository,
-				"trivy.useBuiltinRegoPolicies":    "true",
-				"trivy.supportedConfigAuditKinds": trivy.SupportedConfigAuditKinds,
-				"trivy.resources.requests.cpu":    "100m",
-				"trivy.resources.requests.memory": "100M",
-				"trivy.resources.limits.cpu":      "500m",
-				"trivy.resources.limits.memory":   "500M",
-			},
-		}, cm)
-	})
-
-	t.Run("Should not overwrite existing config", func(t *testing.T) {
-		testClient := fake.NewClientBuilder().WithObjects(
-			&corev1.ConfigMap{
-				TypeMeta: metav1.TypeMeta{
-					APIVersion: "v1",
-					Kind:       "ConfigMap",
-				},
-				ObjectMeta: metav1.ObjectMeta{
-					Name:            "trivy-operator-trivy-config",
-					Namespace:       "trivyoperator-ns",
-					ResourceVersion: "1",
-				},
-				Data: map[string]string{
-					"trivy.repository": "gcr.io/aquasecurity/trivy",
-					"trivy.tag":        "0.35.0",
-					"trivy.severity":   trivy.DefaultSeverity,
-					"trivy.mode":       string(trivy.Standalone),
-				},
-			}).Build()
-		resolver := kube.NewObjectResolver(testClient, &kube.CompatibleObjectMapper{})
-		instance := trivy.NewPlugin(fixedClock, ext.NewSimpleIDGenerator(), &resolver)
-
-		pluginContext := trivyoperator.NewPluginContext().
-			WithName(trivy.Plugin).
-			WithNamespace("trivyoperator-ns").
-			WithServiceAccountName("trivyoperator-sa").
-			WithClient(testClient).
-			Get()
-		err := instance.Init(pluginContext)
-		require.NoError(t, err)
-
-		var cm corev1.ConfigMap
-		err = testClient.Get(context.Background(), types.NamespacedName{
-			Namespace: "trivyoperator-ns",
-			Name:      "trivy-operator-trivy-config",
-		}, &cm)
-		require.NoError(t, err)
-		assert.Equal(t, corev1.ConfigMap{
-			TypeMeta: metav1.TypeMeta{
-				APIVersion: "v1",
-				Kind:       "ConfigMap",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:            "trivy-operator-trivy-config",
-				Namespace:       "trivyoperator-ns",
-				ResourceVersion: "1",
-			},
-			Data: map[string]string{
-				"trivy.repository": "gcr.io/aquasecurity/trivy",
-				"trivy.tag":        "0.35.0",
-				"trivy.severity":   trivy.DefaultSeverity,
-				"trivy.mode":       string(trivy.Standalone),
-			},
-		}, cm)
-	})
-}
 
 func TestPlugin_GetScanJobSpec(t *testing.T) {
 
@@ -6879,7 +6095,9 @@ default ignore = false`,
 				WithClient(fakeclient).
 				Get()
 			resolver := kube.NewObjectResolver(fakeclient, &kube.CompatibleObjectMapper{})
-			instance := trivy.NewPlugin(fixedClock, ext.NewSimpleIDGenerator(), &resolver)
+			psm, err := trivy.NewPodSpecMgr(pluginContext)
+			assert.NoError(t, err)
+			instance := trivy.NewPlugin(fixedClock, ext.NewSimpleIDGenerator(), &resolver, psm)
 			securityContext := &corev1.SecurityContext{
 				Privileged:               pointer.Bool(false),
 				AllowPrivilegeEscalation: pointer.Bool(false),
@@ -7289,7 +6507,9 @@ default ignore = false`,
 				WithTrivyOperatorConfig(tc.trivyOperatorConfig).
 				Get()
 			resolver := kube.NewObjectResolver(fakeclient, &kube.CompatibleObjectMapper{})
-			instance := trivy.NewPlugin(fixedClock, ext.NewSimpleIDGenerator(), &resolver)
+			psm, err := trivy.NewPodSpecMgr(pluginContext)
+			assert.NoError(t, err)
+			instance := trivy.NewPlugin(fixedClock, ext.NewSimpleIDGenerator(), &resolver, psm)
 			securityContext := &corev1.SecurityContext{
 				Privileged:               pointer.Bool(false),
 				AllowPrivilegeEscalation: pointer.Bool(false),
@@ -7471,6 +6691,8 @@ func TestPlugin_ParseReportData(t *testing.T) {
 		Data: map[string]string{
 			"trivy.repository": "aquasec/trivy",
 			"trivy.tag":        "0.9.1",
+			"trivy.mode":       string(trivy.Standalone),
+			"trivy.command":    string(trivy.Image),
 		},
 	}
 
@@ -7550,9 +6772,10 @@ func TestPlugin_ParseReportData(t *testing.T) {
 					"generateSbomEnabled":  "false",
 				}).
 				Get()
-
 			resolver := kube.NewObjectResolver(fakeClient, &kube.CompatibleObjectMapper{})
-			instance := trivy.NewPlugin(fixedClock, ext.NewSimpleIDGenerator(), &resolver)
+			psm, err := trivy.NewPodSpecMgr(ctx)
+			assert.NoError(t, err)
+			instance := trivy.NewPlugin(fixedClock, ext.NewSimpleIDGenerator(), &resolver, psm)
 			vulnReport, secretReport, _, err := instance.ParseReportData(ctx, tc.imageRef, io.NopCloser(strings.NewReader(tc.input)))
 			switch {
 			case tc.expectedError == nil:
@@ -7834,7 +7057,9 @@ func TestGetContainers(t *testing.T) {
 				WithTrivyOperatorConfig(map[string]string{trivyoperator.KeyVulnerabilityScansInSameNamespace: "true"}).
 				Get()
 			resolver := kube.NewObjectResolver(fakeclient, &kube.CompatibleObjectMapper{})
-			instance := trivy.NewPlugin(fixedClock, ext.NewSimpleIDGenerator(), &resolver)
+			psm, err := trivy.NewPodSpecMgr(pluginContext)
+			assert.NoError(t, err)
+			instance := trivy.NewPlugin(fixedClock, ext.NewSimpleIDGenerator(), &resolver, psm)
 			jobSpec, _, err := instance.GetScanJobSpec(pluginContext, workloadSpec, nil, nil)
 			assert.NoError(t, err)
